[roll] Roll fuchsia [fidlc] Add top level raw::TypeConstructor
In order to support raw AST nodes that can be reused
across syntaxes, this CL introduces a top level TypeConstructor
type that abstracts away the fact that the underlying class
may be TypeConstructorOld or TypeConstructorNew. This
approach is preferred to simply introducing _Old and _New
copies of each node that we want to reuse in both syntaxes,
which involves a lot copied code.

The implementation accomplishes this by using std::variant -
an alternative implementation using inheritance is explored
in PS3. std::variant is preferred because the "branching"
is located at the call site rather than at the definition site,
which is good in this case since we don't necessarily want
code in raw:: to know about the callers elsewhere. Another
minor benefit is that the specific new/old implementations
can have different names/type signatures, which is not the
case when using dynamic dispatch.

As an example, this new top level TypeConstructor type is
used in raw::AliasDeclaration (which previously used a
Kind enum and mutually exclusive fields). The use site
code can be seen in flat::Library::ConsumeTypeConstructor,
raw::File::ParseTypeConstructor, and
TreeVisitor::OnTypeConstructor.

Once this CL lands, other "dual use" raw AST nodes will
be updated to use TypeConstructor as well.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507279
Original-Revision: c81ce025b0819851bd54718b7cd0d91e152c9469
GitOrigin-RevId: e2f9f31534820f70f5f2ee9780239648dbf83a6b
Change-Id: I97bebd18246748856e8dc9528a04118165229ced