[roll] Roll fuchsia [kernel][phys] Ensure %gs is set up to point to beginning of TLS area.
In phys _start, we currently point %gs to the end of the TLS section,
instead of the beginning as required.

In particular, x86-64 uses the offsets:

  #define ZX_TLS_STACK_GUARD_OFFSET 0x10
  #define ZX_TLS_UNSAFE_SP_OFFSET 0x18

for TLS, both of which require %gs to point to the beginning of the
TLS area, not the end.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508578
Original-Revision: c09a549e2c99ab624fd7e06dbf6f23cc78ba1252
GitOrigin-RevId: 1055f44ce0dde463ab41e63cbebf0f3256df7a04
Change-Id: I911adf72352d991f243d7fa4fba0eb51d7c31751