[roll] Roll fuchsia [zbitl] Have iterator hold the current item's offset
Currently it holds that of the next item. This means that the
determination of item and payload offsets is done by working backwards,
which assumes that headers and payloads are alternatingly laid out and
that headers are of a fixed length; support for the bootfs case will
invalidate these assumptions.

Original-Bug: 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508239
Original-Revision: 2d6db603e61f9e6f7f89cfa486dde1b8283f3f38
GitOrigin-RevId: ee8aa0a681a0c2b2ea56a5dbf995f823177efda7
Change-Id: I55e639ce05fbaedd1e8353d372f685659fbb9f2d