[roll] Roll fuchsia [pkg-cache] Allow unexpected blob closes
Relax the NeededBlobs protocol requirements a bit to allow callers to
close a blob while writing it and later re-open the blob for write, and
fill in some missing test coverage for this case.

The package resolver handles blob fetch errors by closing the blob and
starting over from scratch, and the pkg-cache should continue to allow
that behavior when the pkg-resolver switches to using pkg-cache.

Original-Bug: 64627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508356
Original-Revision: 5f5c71394918bfd801862ca8d098972f95b10612
GitOrigin-RevId: 0c49de13e972f64c94b9c9a5418d4ab3a8931836
Change-Id: Iac56f358250599fbbbbf6c98ddf012420a8c60a5