[roll] Roll fuchsia [pci] Enhanced legacy interrupt support
This change contains the majority of the implementation for RFC-0070.

When a userspace PCI_IRQ_MODE_LEGACY IRQ is triggered the bus driver's
IRQ worker will now disable that driver's IRQ generation ability until
the interrupt is handled and acked. This is functionally similar to how
the kernel PCI driver handles this, except it must be done in userspace.

A new legacy IRQ mode, PCI_IRQ_MODE_LEGACY_NOACK, has been added. In
this mode the bus driver will allow up to a configured number of
interrupts per second from a given device before determining a device
has too many spurious IRQs. In the event that this happens the device's
IRQ generation will be disabled until the device changes IRQ modes. In
this CL that limit is 512 interrupts a second, or an interrupt every 2
milliseconds. This presently is set at bus level, but can be both
adjusted and moved to per-device constraints in the future if testing
finds it necessary. Since high performance devices will be using MSI
or MSI-X typically, we may find this limit to be acceptable in practice
provided the bus driver's CPU usage stays low.

Original-Bug: 32978, 71188
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472848
Original-Revision: b19112d265ccfc0738d8d09a661a744f3a992eb1
GitOrigin-RevId: 7c80781da7f6dd83c4041e902acce498bc09d019
Change-Id: I87bd66c3450b5144e55977116677f5d224b9d3f6