[roll] Roll fuchsia [io2] Make fuchsia.io.AddInotifyFilter match fuchsia.io2.AddInotifyFilter
This changes the path from a string to path-specific alias defined in
io2.

The string limit is one shorter, noted as desirable at
https://cs.opensource.google/fuchsia/fuchsia/+/master:sdk/fidl/fuchsia.io/io.fidl;l=288
so this is strictly wire-incompatible, but seems unlikely to cause a
problem in practice.

Source-wise, the bindings treat the path as either a string view, or a
string, not a fixed sized buffer, so also shouldn't be a problem.

Additionally, copy reformatted comment from io2.

Original-Bug: 71365
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504531
Original-Revision: 4e543996b2b77eee0b971d5e3a566796a58ac0e6
GitOrigin-RevId: f07e28e87c5d53be171246c56f96dc4a918a0b8d
Change-Id: I9c7627d1f0ebc76ceaa1656ea94c57e39023beae