[roll] Roll fuchsia [bt][a2dp] Fix cobalt metrics reporting
Metrics reporting often is failing because we are not detaching the
task, and when the Media task is stopped by the stream it is often just
dropped.

Detach the cobalt reporting for now to make sure it submits it's report
to the Cobalt server.

Original-Bug: 71825
Test: added dropped_task_reports_metrics

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508879
Original-Revision: 0999e39106dd613fed2a8d617cb243e7985eea6c
GitOrigin-RevId: 772835713f35f8767db2003740b8540be52eb4fa
Change-Id: I61c7153d8563740f44a3455516996031f13cd633