[roll] Roll fuchsia [kernel][hypervisor] Fix EL2 exception vector tables alignment
Sections D12.2.(116-118) (VBAR_EL[123]) of the ARM ARM show bits [0,
10] of the base address being reserved and must be set to 0.  The
implication here is that the exception vector table must be aligned on
11 bit boundaries.

Unfortunately, until now, we have been aligning them on 10 bit
boundaries, not 11.  This has not manifested as a problem yet for a
couple of reasons.

First, at the top of el2.S, there is a directive to align to a page
boundary (12 bits), followed by a number of macros.  The first actual
generated code is one of the two exception vector tables which then
specifies a less strict .align 10 directive, when the generated code
is already aligned to a 12 bit boundary.

If _any_ code is added before the table, however, this gets undone.
The code pushes the table forward, and since it was already 12 bit
aligned, it become 10 bit aligned, on an odd boundary (eg, bit 10 of
the address is 1, not 0).  When an exception from EL1 is taken, it now
hits the start of the table, which was meant to handle EL2 -> EL2,
SP0 exceptions (which is currently an invalid exception handler).

This leads to the second reason that this is easy to miss.  QEMU
completely ignores this requirement.  10 bit alignment is just fine by
it, so things continue to work.  If you switch to testing on an Astro,
however, you can program the address, and even read it back as what
you wrote, but the bottom 11 bits are always ignored.

At any rate, fix this by changing the alignment directives in front of
each of the tables, so that if someone adds any code in front of the
tables in the future, they don't get hit by this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509467
Original-Revision: d2a4401611663abad195859cef75b15f0b6fb5f2
GitOrigin-RevId: 6a51444dd3afbbe309d7ca165b12849adea6bfd7
Change-Id: I734c986220d5ae79f0e5a2dc4c5ef9861e02686a