[roll] Roll fuchsia [brcmfmac] Avoid debug asserts on rxglom failure
If RX glomming failed to parse the received header, because of a
transfer error or something else, the code would attempt to use
uninitialized values from that parse. This was harmless in release
builds because the computed values were never used but in debug builds
they trigger asserts that crash the driver.

Check the return code of the header parsing and only do the computations
if the parsing succeeded.

Original-Bug: 72260
Test: Not able to reproduce CRC error but verified normal operations
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508838
Original-Revision: fab3dd7e20547122f038f1ce279b2b7d4e5b8912
GitOrigin-RevId: 664f2b17df489ef8ca8ad10c2ac77ad93928dd04
Change-Id: I1c3a11bded5de5de2e634b774c50977eed15b3d9