[roll] Roll fuchsia [volume_image]: Introduce command adapters.
Each of these commands will map to an existing legacyy tool command and
options. Will make testing the behavior easier. Later will introduce
argument to parameter conversion, which can also be tested,
which will make the binary a trivial call to parse options and command
and a call into the command adapter.

Tests are intentional superficial, in the sense that the library logic
is tested, but just checking that the options are passed.

To use fvm-check, wrapped in if def a section that tries to map to some
fuchsia only code, to figure out the name of the partition. This allows
the tests to run in host and fuchsia.

Test: storage-volume-image-test(host),
storage-volume-image-tests(fuchsia),
storage-volume-image-adapter-integration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507783
Original-Revision: 387f7a8146a5dedc2a2a49908a07d63c8e9decaa
GitOrigin-RevId: c4b1c7ba79b9f737227599e943fa6733aa002e2f
Change-Id: I981fe47d77a4b18a6d10c0e79781fd8e21f49c5b