[roll] Roll fuchsia [starnix] Generate uapi using bindgen
This CL uses bindgen to automatically generate some of the constants and
types from uapi that we had previously typed in manually.

Currently, we need to run the bindgen tool manually using a script.
Hopefully we'll be able to run bindgen as part of the build in the
future.

We seem to be using a different signed/unsigned type for various
constants than what bindgen generates. In this CL, I've left put in a
manual conversion in types.rs, but we should study whether we can use
the signed-ness that bindgen generates without needing the conversion in
a future CL.

Original-Bug: 70902
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509810
Original-Revision: 9c194a1a5a4322f2b7633c4a7f7c1685afba04a3
GitOrigin-RevId: a98c105a6c2c4a6f7b6ee7a5c99bad390b8282b5
Change-Id: Ia72012f9b152db3580accab72bde7365fdd86fbc