[roll] Roll fuchsia [fidlgen][cpp] Migrate fuzzer generators to use CodegenOptions
The fuzzer bindings have so far hardcoded the include stems of the C++
bindings which it depends on. This CL changes it to instead take the
include stems from the command line.

A nice effect is that all include stem source-of-truths are now defined
in fidl_cpp.gni, making it easier for us to change them later.

We are also able to remove the FuzzerHeaders field in the IR, which used
to simply duplicate Headers. This was because the fuzzer codegen was
overwriting Headers with its own values to point to the HLCPP/LLCPP
bindings. After making the binding include stem first-class, we are able
to simply use Headers to represent fuzzer headers, same as all other
binding flavors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508542
Original-Revision: b081228b79236be4f425766732619af35a95417b
GitOrigin-RevId: 49fb6c4e0fd578999e32cec0fcbefd2497b417a0
Change-Id: I20f1972ee664029949751634739f04aab102c5c4