[roll] Roll fuchsia [Bind] Fix symbol encoding in new bytecode
Modify the mismatch symbol type check in the new
bytecode encoding. Since the LHS symbol represents
the property key and the RHS symbol represents the
property value, the symbol types might not match.

Instead, the code now checks if the LHS symbol
represents a key and the RHS symbol represents a
value. If the LHS symbol contains a value type,
it is compared with the RHS symbol to ensure that
they match.

Original-Bug: 71051
Test: Updated and added unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505938
Original-Revision: 1116506101c02a7e27e92f81eda86e92c5d7f6e9
GitOrigin-RevId: bc3e552471461bbf136054cbdb073918aff03034
Change-Id: Ice5ffff3e11185a04557dc3ab0def254fbd3f984