[roll] Roll fuchsia [inotify] Add localfileystem,remote-io, zxio support.
This CL makes changes to remote/zxio layers as well
as local-filesystem layers to make the FIDL call to
VFS via DirectoryIo protocol.

Also changed the ordering of the arguments to be consistent
throughout the various calls to AddInotifyFilter.

Also, we change zx::socket to zx_handle_t to pass via
the zxio layer which needs to be compatible with the
C interface and then later wrap it up back into
zx::socket for the fidl calls.

The readv method is intentionally left unimplemented,
and will be done as a followup CL with tests.

Testing: Added fake server to emulate DirectoryIO protocol
         to test if we are calling the FIDL correctly from the
         namespace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488778
Original-Revision: 8257d14097249188ad816d2d7e35191c4575c17c
GitOrigin-RevId: 48c1aec33d3580e7b85fdc2e52ee039e4ec9cd68
Change-Id: Iaae759e75af4f5d56fb824da469dcd01e6805875