[roll] Roll fuchsia [session] Don't connect to Launcher in element management tests
SimpleElementManager::new() connects to to the Launcher protocol
in the component's incoming directory. This protocol is not routed
in tests (and it shouldn't be), resulting in logspam, e.g.

    [00496.513644][108203][108380][<root>] WARN: No capability
    available at path /svc/fuchsia.sys.Launcher for component
    /tests:auto-9262243642210911116:24/test_wrapper:0/test_root:0,
    verify the component has the proper `use` declaration.

This also causes tests that use `run_until_stalled` to flake if
the test body finishes before the connection fails:

  [00496.513405][1195728][1195730][<root>][element_manager_tests.cm]
    INFO: launch_v2_element(...)
  [00496.513561] ... ERROR: PANIC: Stalled without completing. ...
  [00496.513644] ... WARN: No capability available at path ...

Original-Fixed: 73241
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509143
Original-Revision: 8dcd89206f5a81451fae3334dd154ca2ba2c246e
GitOrigin-RevId: cf26b2f93e5103eba57f1e4ba0e5a3de9f388646
Change-Id: I378c8f8ccde7e688b906a3efe4ffb7976051b81a