[roll] Roll fuchsia [bt][snoop] Use monotonic clock when UTC is not available.
Store all packet timestamps in clock monotonic time. Apply
a transformation to UTC when serializing packets to FIDL or
Inspect. Keep timestamps in clock monotonic, if UTC is not
available when serializing.

FIDL clients that are already following the packet logs when
the UTC clock becomes available will continue using clock
monotonic for the remainder of the session. This is done to
avoid a discontinuity in timestamps.

Original-Bug: 72835
Test: unittests for clock.
Test: manually checked wireshark output is correct.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507782
Original-Revision: 21cda78d7de8a4320b50a2e3074ddf559a63fb15
GitOrigin-RevId: 88bbc4cf5e10210a5a7220785c5ac493e584f15f
Change-Id: I5d3ab105cc1a5536e39c69b576453759a50ea2ca