[roll] Roll fuchsia [sdk][cts] Add "cts" Category of SDKs
This change adds a "cts" category of SDKs. This enables the creation and
release of a CTS-specific SDK, which can be used as a mechanism to
release CTS sources. A CTS-specific SDK would be used to release
libraries required to build the CTS, but should not be released /
depended on in a public/partner SDK.

Original-Bug: http://fxbug.dev/72547
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506838
Original-Revision: 3630e928c49e686d99246fef4272d0f7746939da
GitOrigin-RevId: 1dd79dc5fe17371d04dc4f687d424763780e7ac5
Change-Id: I93d966ed57121f0f0dfe75201f77a29f40b9e4b6