[roll] Roll fuchsia [wlan][sl4f] always set client controller
Previously, if the facade already had a client controller when
create_client_controller was called, the facade would not replace the
client controller. However, if something has the client controller when
create_client_controller is called, we may not get an error until we try
to use the facade's client controller. The previous implementation would
not try to create a new client controller if the facade already has one,
so there would be no way to get a working controller in this situation.

This implementation will always drop the current controller and then
create a new controller.

Original-Bug:71244
Test: Ran a policy ACTS test and verified that it still works:
	act.py -c <path to config> -tc SavedNetworksTest
      Manually tested that calling create_client_controller twice in an
      ACTS test would not cause an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509473
Original-Revision: 6f733d70cd977b63e15ed1e71d4bb1b35ea80c86
GitOrigin-RevId: b4c5ae5397f71ceae1104e106e02796c38a7ce1d
Change-Id: I88e2f4e255f152673649cbf4dbc76b0744857e03