[roll] Roll fuchsia [fidlgen][cpp] Helper for doc comments
Basically, replace:
  {{- range .DocComments }}
    ///{{ . }}
  {{- end }}
with:
  {{- .Docs }}
in C++ bindings.

This makes the templates have less visual and semantic noise. Perhaps
most importantly it removes a scope so there's one less {{ end }} to
keep track of.

To do this we have to be consistent on how we put doc comments into C++
and I went with triple-slash as was used in the HLCPP comments because
it's supported by Doxygen.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510083
Original-Revision: 5b39da399585e4a3a712262cca50eac73e9be2e5
GitOrigin-RevId: 8f2d50ba63df405a341e9d38294be48b0361f7c2
Change-Id: I5787244819686dbd148ce3852eb4a15914b964af