[roll] Roll fuchsia [ffx] Correctly get tools from modular SDK in tree
Moving to the modular SDK will put the in-tree manifest in a new
arch-specific location. The upside of that will be only one of each tool
in the manifest, and selecting the right manifest means always getting
the tool with the right arch.

Right now building the modular manifest is optional, but to be ready for
the transition, we always look for it and fall back to the old location
if it's missing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506865
Original-Revision: effeee8af6673b8bb1cbea91f3c9dda230f4133b
GitOrigin-RevId: 8cc05675cf40753f11d49d76ad098b18f5a9dae5
Change-Id: I2aac1b7d785ccdfdec23f01b62f57b5895b91a92