[roll] Roll fuchsia [zircon][kernel] Support for ZX_INFO_VMO_IMMUTABLE
This CL is an implementation of RFC-0080 and adds support for
determining if a VMO child was created in an immutable state (and will
remain immutable).

For this to be the case, the VMO clone must be created with the
ZX_VMO_CHILD_NO_WRITE and ZX_VMO_CHILD_SNAPSHOT flags and without the
ZX_VMO_CHILD_RESIZABLE flag. The child VMO also must not be discardable,
but there currently is no way to create a discardable child (and
therefore also no way to test for this case).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/508658
Original-Revision: 0e83c15f5d0f7408f441bd1319b70f187f82bf42
GitOrigin-RevId: c8250f0d67c9e59975c096dbeb9a28e333b0a06e
Change-Id: I0dd8324b907b7d29f1f9e0d52aa659a37f93c66a