[roll] Roll fuchsia [session] Add scene manager component
This component extracts scene management and input code out of the
workstation session. This first step keeps all the functionality as
similar as possible, and introduces a simple FIDL protocol that
workstation can speak to the scene manager to configure the scene and
input pipeline.

This FIDL protocol will need to be updated prior to inclusion in the
SDK, as it currently contains a legacy input event listener.

Ideally this component will, after removing some of the workstation
specific quirks, be able to be used across products to set up the scene
graph and handle input.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487757
Original-Revision: bd0690285627757247d0a046d0316011828c65d1
GitOrigin-RevId: d14a122bc8fd0ab8402394ffa6e909ab567bf3ad
Change-Id: I3cf970fbfbb6518837887182368ec9e30f9e196f