[roll] Roll fuchsia [testsharder] Run multiplied test at least once.
There was a bug where testsharder would calculate the number of runs for
a test (if the total_runs was unset) by dividing the target shard
duration by the test duration. If the target duration was less than the
test duration, than the number of runs calculated was 0 so it would
assign 0 tests for that multiplied shard to run. This was written to the
json output file as `null` which would give a type error when processed
by the recipes. We should instead run it once.

Original-Bug: 73515
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510087
Original-Revision: fa6b2a6bf8f9152622e59559e6b4834e47d58bdf
GitOrigin-RevId: c2c5b91b181c5d32bfc2efdde9aa2780bde33677
Change-Id: I409b4d1099da6ec87bae20ec4ae9960ef196f568