[roll] Roll fuchsia [modular] Refactor annotation management
This updates annotations to be solely owned by element manager and
allows the presenter to read and update them.

The presenter is no longer notified of updates to annotations
through the ViewController. This will be added to AnnotationController
directly in fxbug.dev/56594.

Test: fx test modular_lib_tests && fx test modular_integration_tests
Original-Fixed: 67592

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509042
Original-Revision: 9aeac89dea9381487b25d77894e5b5e7195a2d89
GitOrigin-RevId: 425f0e466025e6dc48105fe83db7eb978ce3346d
Change-Id: I5508ff3357a29441fedf907c1c9d1e555e70a51a