[roll] Roll fuchsia [kernel][asan][arm64] Use new ArchAspace::Map interface.
The new interface for mapping pages requires the caller to specify an
action to perform when a mapping already exists in place.

We are using the same behavior as before: fail. The code in kasan
already checks whether a mapping exists before doing it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510321
Original-Revision: fa6523d7ad566f5c8f441943e1ce5a0799f27152
GitOrigin-RevId: 4cc07cdb54e03e940f1d5832d893ec48168be821
Change-Id: Icfbe69d7af8a9f271c0b38d2659d7daf80f94992