[roll] Roll fuchsia [bt][at] Add support for optional arg delimiters.
Typically, AT command arguments are offset from the command name with an
=.  Sometimes, however, they can be offset with > or no character at
all.

The first of these cases was already supported as a "nonstandard
delimter" in the AT command library.  Adding support for the second
required some restructuring of the Definition and Command types.  The
optional nonstandard delimiter is changed in this commit to an optional
delimiter string.

This introduced some issues in parsing, as the parser can now see the
lack of arguments to a command as an empty delimiter, followed by an
empty argument list.  To remove this ambiguity, I made execute commands
arguments mandatory.

This necessitates changes to tests and generated translation code.

Original-Bug: fxbug.dev/66041 fxbug.dev/72873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505561
Original-Revision: 4195d239e28b1b3a31b41559012282d4ffe72307
GitOrigin-RevId: e14b08e03e66e9ee94b5d6b07936afde3af824f3
Change-Id: Id3928952e61a627ebc6fadef25154abfa41d0cbc