[roll] Roll fuchsia [fidlc] Implement parsing of layout subtypes.
Value layouts `enum` and `bits` previously did not support the parsing
of "wrapped type" in the new syntax.  The code to parse these tokens is
a bit tricky, as the parser cannot unambiguously decide if a colon that
appears like `enum : FOO` denotes the start of a constraints block, or a
wrapped type.  This is resolved by parsing `FOO` as a constant, then
checking the next token to see if its a `{` - if it is, we assume that
`FOO` is a wrapped type, otherwise, we assume it is a constraint.

Test: fidldev test fidlc
Original-Bug: 65978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509805
Original-Revision: e27dca7e8009e2124dd0aac95e4e8a6a90ca1cff
GitOrigin-RevId: 1a36dabb0b935ae9fb8177b388e86a401c6b9375
Change-Id: I862ff0dbea8df73e1d5829a69e4280ce779d4459