[roll] Roll fuchsia [zbitl] Abstract ZBI-specific logic in View
Soon we will move the core of View's container-y business logic into a
subclass templated on "container traits" (which will too will be used
for a BOOTFS analogue). To ease that transition, we introduce the
types, and static members and methods that will comprise the traits of
the ZBI format.

`header_type` was renamed to `item_header_wrapper` to avoid confusion
with `item_header_type` and `container_header_type`, both introduced in
this change.

Container header logic is updated to no longer takes advantage of the
optimization provided by header_type/item_header_wrapper, as BOOTFS
container and item headers are distinct. The cost of this, however, is
negligible: it amounts to at most a handful of bytes copied over the
lifetime of a View.

Original-Bug: 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490297
Original-Revision: ffc658804e36174e7abfa2475e651587eaffe5e2
GitOrigin-RevId: d7140a9ec3d88f6af0d82b205b5319244e399827
Change-Id: I7f6e1e52d48757d51fa118ac2c36ba2960e68c9a