[roll] Roll fuchsia [carnelian] Don’t render when display isn’t owned
The main problem with this approach is that when ownership of the
display is returned, an incorrect pre-rendered image is shown. I’d like
to go with this now, though, and take more time to consider how to make
sure the first frame on regaining ownership is correct.

I noticed that the fuchsia-framebuffer example box was a little broken
so I fixed it and updated the manifest file for the integration test.

Original-Bug: 73628

Testing: manual, by checking that spinning square doesn’t appear in top
when it doesn’t own the display.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510582
Original-Revision: d1696d22eea18829e30e9efddae0f2fc5d7cc9c9
GitOrigin-RevId: c2a24570a3a92f310435376f45c960eef241e6a4
Change-Id: I7fdcee2a686d3f6278667430f5b962c42bc936b3