[roll] Roll fuchsia [camera] Handle client token failures async
Instead of using ValidateBufferCollectionToken to see if a token is
valid, register an error handler and call Sync. This eliminates the need
for the client to call Sync before using SetBufferCollection and also
improves handling for the error case of a token server end being closed.

Also ensure that a client provided token is not included in other client
buffer requests until a Sync has completed to validate there's a
responder on the other end.

Adjust unit test to test for the case of a token failing after being
passed to SetBufferCollection.

Original-Bug: crbug/1192720
Test: camera_device_unittest
Multiply: camera_device_unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509044
Original-Revision: d91f063818ecfc32633c95c1e512dc72b946775b
GitOrigin-RevId: 6ce439ac3eed99919e585c79cf99d2a67c13186a
Change-Id: Ib55143ac908e9e26f3a46c6be315e93e72218853