[roll] Roll fuchsia [covargs] Run llvm-cov on individual binaries to check for validity.
Some of the Rust test binaries are giving a `Malformed instrumentation
profile data` error. This should be looked into but in the meantime we
should skip these files so that we can still generated the coverage
report with the rest of the valid profile data.

Original-Bug: 27171
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511398
Original-Revision: 7c8f1a402e24a89cbde32f80486fcd05bc326107
GitOrigin-RevId: 8b8a827c67a8fd3defa417cbbe953574482328f1
Change-Id: Ia902d63ffb56c9342b6dcd0a5af403fb2bb908ee