[roll] Roll fuchsia [audio_core][tests] Fix AudioCoreHardwareTests
This test-only CL fixes sources of intermittent results in the
audio_core_hardware_tests suite.

The system versions of AudioCore and AudioDeviceEnumerator are
used, instead of hermetic versions that would compete with the
system instances for access to audio drivers and devices.
AudioDeviceEnumerator is added to the kAllowedSystemServices
list, because these tests rely on actual hardware in a way that
cannot be faked (we use the analog noise floor as proof that we
have captured an actual live input stream, not than just a
zeroed-out sensor).

The retrieval of a purportedly "preferred" format is removed,
using instead the very-common 2-channel 48kHz configuration.

Default input device is tracked with an optional, clarifying
whether it has been set.

Device arrivals/departures/default_changes are not automatic
failures until right before we set gain levels and start
capturing. During this interval, we discern between default
changes of output versus input devices. If the default input
device is removed (such as the VAD, used in a non-hermetic
end-to-end test), then we will use the new default, but we
require that there be an input device default by the time we
set gains and start capturing.

The test setup sequence is adjusted; we set the format and
submit the payload buffer before connecting a GainControl and
normalizing the loudness. (This is just being fastidious; it
was not considered a source of flakiness.)

In addition to a simple check of whether ANY captured sample
is non-zero, we now also check for detectable DC offset, by
comparing the (absolute) mean to the standard deviation.
Assuming the non-zero-sample test is useful, this second test
ensures that the test devices are not miscalibrated in ways
that cause false-negative passing results.

Multiply: audio_core_hardware_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511559
Original-Revision: ee5a3a11adcd764b4fc1400dd704da7f775d5d45
GitOrigin-RevId: 07e202c5041ce05e18f609e84255af43bd880d0d
Change-Id: I97d24042a75bd48f1830783f8e608b2e8a20ea54