[roll] Roll infra/recipes [prebuilt_host_tool] Support uploading to GCS
The recipe needs to be able to build ffx and upload it to a GCS bucket,
and it's easy enough to make this generic rather than hacking in
something that's specific to ffx.

This involves extending the properties schema to let builders specify
whether each tool should be uploaded to CIPD vs. GCS, and the
destination (CIPD package name or GCS bucket/path).

This is a backwards-incompatible change because it changes the schema of
the `tools` field, which will require a corresponding change to the
builders that use this recipe. It's very tricky to do a soft transition
for this because of the type strictness of protobufs, so a soft
transition would require a complicated double field renaming. The
builders that use this recipe are low-priority anyway (they were
recently broken for 4 months and no one noticed) so I figure it's not
the end of the world if they fail once or twice between landing this
change and landing the corresponding config change.

Original-Bug: 72680
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/505759
Original-Revision: 3d228c25fc49ed6f52ed5c1ac7c72db535cad21e
GitOrigin-RevId: 32dd8cca20e7f00e03292b65aa4a90c39d09028c
Change-Id: I0cb04d3201b83479ab51d17890c4ab222ccd9040