[roll] Roll fuchsia [fint] Generate ninja graph and compdb
Currently this is done in recipes, but once `fint build` is done we'd
like for the recipes to no longer know how to run Ninja at all. So it
makes sense for `fint build` to handle these.

We pass the paths to the generated files via output build artifacts
message so that the recipe doesn't need to hardcode constant paths to
these files.

Also emit the path to the main ninja log file, as the infrastructure
needs access to that file to do some performance result processing. The
path is constant with respect to the build directory so theoretically it
could be hardcoded in the recipes, but that path is more of an
implementation detail of the build system that falls solidly on the
platform side of the platform-infra interface, so it's best to not
hardcode that sort of knowledge in the recipes.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511441
Original-Revision: 7d35b5f2927105e07abdad087fb1eccd7045cc8a
GitOrigin-RevId: 5944a6cfd5c7956e388b29250541a5532687401d
Change-Id: I849e38c69fe46e4f20bf8859611babfef9826c02