[roll] Roll infra/recipes [fint] Generate ninja graph and compdb
Currently this is done in recipes, but once `fint build` is done we'd
like for the recipes to no longer know how to run Ninja at all. So it
makes sense for `fint build` to handle these.

We pass the paths to the generated files via output build artifacts
message so that the recipe doesn't need to hardcode constant paths to
these files.

Also emit the path to the main ninja log file, as the infrastructure
needs access to that file to do some performance result processing. The
path is constant with respect to the build directory so theoretically it
could be hardcoded in the recipes, but that path is more of an
implementation detail of the build system that falls solidly on the
platform side of the platform-infra interface, so it's best to not
hardcode that sort of knowledge in the recipes.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511441
GitOrigin-RevId: 376460269a692a3556ccffbbf055921a0fd0ec58
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/511858
Original-Revision: 2e74d92f76539bb20c6d3fd2789c26e99c1c2e3f
Change-Id: I4afd23ef508913e1aed7c1abf89d043efeb05065