[roll] Roll fuchsia [build] Add template for verifying two files match
Add a new GN template and python script that tests at build time if
two files have the same contents.

This is a utility for verifying that migrations are resulting in the
same contents by allowing an old and new implementation to reside
side-by-side in the tree, and verifying that they are producing the
same results.

This is not meant for "Golden File" verification, but is mechanically
very similar.  This is for comparing two build outputs against one
another.

Original-Bug: 73184
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511638
Original-Revision: 376102f3ee3e06b307efc0cd49ae029888fd6458
GitOrigin-RevId: 3439908abb105450d6febc41561324d9c63f182c
Change-Id: I0546a9fbcfce39c8a9a5974babee5e8f7f6c91e2