[roll] Roll fuchsia [fx][regen-goldens] Add --list and --reset options
This CL adds two options to `fx regen-goldens`: --list prints the
supported tools and their golden directories, and --reset deletes
everything first to help get out of a bad state.

Before, the supported tools were listed in the help message. This was
not ideal because I always had to remind people to update the help
message in addition to the all_tools array:

* https://fuchsia-review.googlesource.com/c/fuchsia/+/506315/3..9/tools/devshell/contrib/regen-goldens#b22
* https://fuchsia-review.googlesource.com/c/fuchsia/+/506782/2..10/tools/devshell/contrib/regen-goldens#b23

Another issue with regen-goldens was failures in GN or ninja at an early
stage preventing regen from completing. This can happen if the
goldens.txt file is missing, or is incorrect, or if golden files cause
formatters to fail (e.g. due to merge conflict markers). With this CL,
regen-goldens suggests you re-run with --reset, which first resets the
goldens directory to a clean slate, avoiding those issues. This causes
the build system to do more work, so it is not the default.

Test: fx self-test subcommands/fx_regen_goldens_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510088
Original-Revision: 35acc60115cbf2faaed68d211053e18556150280
GitOrigin-RevId: be8b5f818ec976e9934c26fc3c7054bf04705173
Change-Id: I61ac3dfd6daa03d5e1c1f1dd5322c6e8d6287fca