[roll] Roll fuchsia [driver-manager] Seperate Matching from Binding
In coordinator we should seperate 'Matching' a Device's
properties to a Driver's Bind code to the act of 'Binding' the
Driver to the actual device.

Once this is seperated, we will be able to keep all of the
Drivers in a seperate DriverMatcher class, which will eventually
be replaced by the DriverIndex component.

Original-Bug: 73827

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510800
Original-Revision: 12b20ce4a626c2342a5a4c875c66d5de5e86aad9
GitOrigin-RevId: 65adf69654148c67bf05604d84e0209c1f96d9c7
Change-Id: Ia50ec4364341081c8cf829621d5f2d60586ca9d4