[roll] Roll fuchsia [build] Add GN arg for enabling/disabling relative-vtables
libneural_network_models.so is currently built using the fuchsia tree
and depends on //build/config/fuchsia/BUILD.gn. This library is used by
a binary built within google3 which exposes a C++ ABI but does not yet
use relative-vtables. In order for this SO to be built without RV, we
will need to make sure that we can enable RV for all of the fuchsia
system without forcing RV on this SO.

This patch allows for RV to be controlled by a build argument
(use_cxx_relative_vtables) which can be set in args.gn. This value will
initially be false, then set to true in a followup one-line patch.

This patch is effectively a no-op.

Original-Bug: 45314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509119
Original-Revision: a9a850d672439fa4d9135b0e7d4d25d94c95e1c4
GitOrigin-RevId: 6899856035648a398f4cd77695c482cd135ccb44
Change-Id: I65452f355893e3a36a9793ddc504968c217cee3b