[roll] Roll fuchsia [netemul] Use test driver pattern
This will allow us to add components such as network_context and
sync_manager as *siblings* of the netemul_sandbox component, as opposed
to being children of it, so that we can route capabilities between them;
for example, the sandbox needs to be able to use
fuchsia.netemul.network/NetworkContext.

This CL also updates the inspect test to expect the 'test_driver' prefix
in the moniker of the components under test.

Original-Bug: 72402, 72403
Test: fx test -d netemul-v2-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510759
Original-Revision: ff2bb6a785dd8465c385e3defb4290bf36d7fa82
GitOrigin-RevId: 0ab60b390dbd0f002e2e1810c28f7bcc818946ff
Change-Id: I4bdb7f31bf1a021177c54593183b8ce0647c3258