[roll] Roll fuchsia [realm_builder] Fix expose bug
If capability route had the source 1/A and the targets 2/B and
2/C RealmBuilder would throw an error because multiple expose
declarations were made from 1 to $root.

This fix checks if adding an expose creates a duplicate in the
vector. Also we add a test to check for this condition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512148
Original-Revision: 5a894b941af982909b76cc2056ca788f2faa2b37
GitOrigin-RevId: c5f1e2d1d39d85efb6416a96dcf1d2099661d06f
Change-Id: I9e4cb6ed3f3bfa10c2a603a88b51782b08902933