[roll] Roll fuchsia [logs] Update text for logs that were rolled out
Currently, we say that logs which were rolled out of the
buffer are "dropped". This is confusing, because it
conflicts with logs which were dropped at the socket level.

This patch updates the display to use the term "rolled out" instead
of "dropped" for these logs.

Original-Fixed: 74001

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512035
Original-Revision: a9c40c41ec3417efdcd8ddd1cffd46b4e359f7c3
GitOrigin-RevId: ae751c3158bfe5fadae0d01ef9520dec2187d3ca
Change-Id: Ibd23867b4d2e9175aeef0ac8927f840293f33987