[roll] Roll fuchsia [runner] Split into separate packages
When the `runner` package was initially created it was intended to hold
implementations of "runners" that abstract away running shell commands
(as local subprocesses, over SSH, etc.). However, it has turned out to
be cumbersome to include widely different runner implementations in the
same package, so we've ended up creating separate packages for different
runners (e.g. `serial` and `sshutil` for running commands over serial
and SSH respectively).

So for consistency, split the `SubprocessRunner` into a separate
`subprocess` package. This also has the advantage of letting us use the
identifier `runner` to refer to subprocess runners without colliding wit
the name of the `runner` package.

At that point `BatchRunner` was the only runner left in the package, so
I moved it into a separate `batch` package. The code is a few years old
and basically un-owned so I didn't want to encourage people to start
using it by keepint it as a library. It will be easy to refactor it into
a library later if necessary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511898
Original-Revision: 46a619dc37211a18dd180ecb50391ed7dd1b80bb
GitOrigin-RevId: 036c1b7e7929bc40fafe89fd845cbedc0a335d54
Change-Id: Ia7fc116f1ef01579ef0732870b3dc64d9ae8d6a2