[roll] Roll fuchsia [block] Make VmoId interiorly mutable
Wraps VmoId's interior ID in an Atomic, so that it can be interiorly
mutated. This is necessary to "take" a VmoId when all we have is a
reference to it.

For example, when doing asynchronous cleanup of some object that owns a
VmoId, the object needs to unregister the VmoId with their block client.
That object would either need to be &mut self in the cleanup context
(which is an onerous requirement), or it would need to wrap VmoId in a
Mutex. All of this can be made easier and simpler by making VmoId
interiorly mutable.

Tested: Built.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509704
Original-Revision: 96c642b1c4ea6cda6f6c77be20c4c25c8b9ca168
GitOrigin-RevId: 8f3d354743d20149add0f82af939a1589d87c9fc
Change-Id: Ia856b27f749545c7773cd1670dd2724d8101028b