[roll] Roll fuchsia [make-fuchsia-vol]: Make explicit the slice size.
Currently relying on the default 8 MB behavior, this was caught during
the migration to a replacement that does not make a default. It is
better to be explicit about the arguments being used, since it makes
things easier to reproduce.

No functional changes.

Test: Builds
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510578
Original-Revision: 7a76bfbfac2aeba9c7b9d1a9ee1f50ec29b004b0
GitOrigin-RevId: 3d4d41fb28a3aed3af976fb7ce45b8f1ba9bfb71
Change-Id: Idba9aea33c2ebdf4aca4c143eb0ff6f9ae413974