[roll] Roll fuchsia [build][ffx] Add GN template for calling ffx from within the build
Add a GN template to correctly call the compiled_action() template
to use ffx as a compiled build action.

This template is a thin wrapper around compiled_action(), and will
automatically add the dependency on the correct ffx build label,
and the binary to the inputs.

The template parses a string of the "plugin name" into individual
tokens for passing to ffx, as well as properly formats a key=value
list of config overrides.

Original-Bug: 73185
Test: local tests using the template
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511679
Original-Revision: 17f6970018d5f141332372a1660f99959b7d579b
GitOrigin-RevId: 1b761a383c7a7785f4b0b874566602471123c40c
Change-Id: I39ebecd8e6cb2c1037856cff555392e331da5df8