[roll] Roll fuchsia [fx] Evaluate $_FX_LOCK_FILE greedily
Instead of evaluating it at the time the `trap` executes. This ensures
that the original file gets cleaned up even if the value of
$_FX_LOCK_FILE is changed for some reason (which should currently never
happen, but we might as well be resilient to that).

This undoes a small portion of fxrev.dev/512098.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513140
Original-Revision: fc2621a7cc5da758be9c67928314288fcc96e749
GitOrigin-RevId: 3d91eeb9b3463bc52e1ad57905509cda853c57d9
Change-Id: I29da5380008bacd0a69f4dc4d3e1ea7b9b73ebde