[roll] Roll fuchsia [kernel][vm] Eliminate warning while reserving range for periphmap
The call to ReserveSpace in reserve_periph_ranges was failing with
ZX_ERR_INVALID_ARGS and generating a "conflicting cache policies"
warning.  Why was it failing with ZX_ERR_INVALID_ARGS?  Because
SetMappingCachePolicy was failing, because LookupLocked was failing,
because it is an error to call LookupLocked on a zero-sized VMO.

In SetMappingCachePolicy, only call LookupLocked (and
arch_clean_invalidate_cache_range) if the VMO has a non-zero size.

In reserve_periph_ranges, ASSERT that the call to Reserve succeeds.

Original-Fixed: 74086
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511633
Original-Revision: 61ea343bf791ca2f0d544cec3e4331934e31387c
GitOrigin-RevId: 2c393255542f3f1d4eb225dc1be68bc690533706
Change-Id: I8ed61bba6e77af2de6a3a38c1480c68667aad8d3