[roll] Roll fuchsia [tefmocheck] Don't attribute logs to tests by default
Many logs are never caused by a specific test, but rather by random
underlying flakiness in the operating system or the infrastructure. It's
not useful to attribute these logs to the tests during which they
happened because there's no pattern to it and it just leads to an
excessive number of flake bugs being filed, all with the same root
cause: for example, 87 flake bugs are currently open for the log
signature ".sock: write: broken pipe", each bug specific to a test:
https://bugs.fuchsia.dev/p/fuchsia/issues/list?q=_broken_pipe%20component%3AFlake&can=2

So only attribute a few specific checks to the test during which the log
occurs.

Original-Bug: 67640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513220
Original-Revision: 11dccf120f199033887bf73300f20d0ef54f6b2b
GitOrigin-RevId: 6bde8d83de8c7ea7ac8d375952130c3e1a53b061
Change-Id: I80c3a1cb971e4f0cf82ca829ed7cf24d1a2c439a