[roll] Roll infra/recipes [build] Use artifacts emitted by `fint build`
The build recipe module currently supports building with `fint build`
instead of `ninja` but still relies on the recipe logic to keep track of
which images and archives were built even though `fint build` now
contains its own logic to emit metadata about the images and archives it
built. This means that we're not yet exercising the end-to-end `fint
build` pipeline, and if we continued migrating to `fint build` with the
recipe code as-is, we'd have to resort to a risky hard transition when
deleting this recipe code after migrating all builders.

To avoid that hard transition, we use the metadata emitted by `fint
build` instead of the metadata assembled by the recipe logic when
running in `fint build` mode. This will ensure that the entirety of
`fint build` will be integration-tested during rollout, and that the
eventual deletion of the recipe code that assembles the metadata will be
a guaranteed no-op on all builders.

Unfortunately, this intermediate state is pretty ugly and required some
duplication of error handling logic. I didn't bother trying to make it
super maintainable because it will hopefully only be around for a couple
of weeks, until all builders are migrated and I delete the old
pure-Ninja code path.

The expectation diffs are mostly a result of:
1. Reordering ninja error handling steps like buildstats and
   clang-crashreports collection
2. Skipping build perf stats generation after running `fint build`, and
   instead using the perf stats generated by `fint build` itself (hence
   slightly different paths for files like the ninja compdb).

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/512740
Original-Revision: 4e3c80d2ecf3012739a576dc5e2d827d69105d80
GitOrigin-RevId: d283b81395af9d99c046296850bedd7268c50e42
Change-Id: I0f09e021fd4e2328e843e23dbfcd4ce6a00b27c3