[roll] Roll fuchsia [volume_image]: Introduce BlockWriter.
There is a custom use case, when paving a device on migration that
writes directly to a block device. This requires aligned writes.

To simplify this, in practice all writes are aligned, or mostly, this
will readback the unaligned blocks, update the the blocks and write them
back. The aligned blocks are passed through.

This should replace the mtd-file in the fvm tool.

Test: storage-volume-image-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512306
Original-Revision: cfe1cd62dfbecdd958cc50be65767ce00a060e1c
GitOrigin-RevId: 1deabd3375f301df7219bd73f3008ec65633e356
Change-Id: If429ff32a5621f3acac167a6446c027e30fbdc1d