[roll] Roll fuchsia [pkg-cache-integration-tests] put isolated devmgr in package
Previously, the tests were failing locally because the devmgr we were
trying to use was in a different package (storage-isolated-devmgr).

I suspect this was working in CQ because that other package was available...
which may not be true locally. Either way, let's fix it to work locally.

Test: fx test -o pkg-cache-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513297
Original-Revision: c525d9392f20fe42ccad1c832fc54d537fc63a07
GitOrigin-RevId: ed7bb629e19ce6ccd7964d4ea0cae09ab2dfa949
Change-Id: I65f7d333688713a174116e571c03ad31e5812889