[roll] Roll fuchsia [netstack] Cleanup expiring DNS server timer
Previously, when an interface was brought down, we did not stop the
timer for expiring DNS servers. When the interface came back up and
rediscovered an expiring DNS server, the server may be prematurely
expired by the stale timer that was not stopped when the interface was
brought down.

Test: fx test netstack-dns-gotests -- -test.run TestTimerCleanup

Original-Fixed: 74201

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513329
Original-Revision: b0d4c0b93af02b4008fda3e0f8e30c243e17e6dc
GitOrigin-RevId: 098e8a0761dc20afb47161f866adb444d60608da
Change-Id: Ied3e2d70af473907784e16808e6376b820f8e1c3