[roll] Roll external/github.com/catapult-project/catapult devil: Add sanity check for root privilege on a device
_TearDownSystemAppModification requires root privilege on a device.
Normally root privilige is enabled on the device before calling
_TearDownSystemAppModification. The function is also re-entrant.
The function may be re-entered after the the device loses root
privilege. For instance if the adb server is restarted before
re-entering the function then the device may lose root privilege.
Therefore we need to do a sanity check for root privilege
on the device and then re-enable root privilege if the device
does not have it.

Original-Bug:chromium:1056214
Original-Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2818568
Original-Revision: ab687ea7be643e0f78176a60f3797aae7fbbae46
GitOrigin-RevId: 0bc58be2e49057752b34a8b509dae3a12bc783ae
Change-Id: If2004cbc2940e54ae0a3f453b8eaadc0f3549362