[roll] Roll fuchsia [fx][qemu] Allow "--no-build" arg to follow other arguments.
Currently, the behaviour of the two command lines:

  fx qemu --no-build -k
  fx qemu -k --no-build

differs. The first runs the default Zircon image in QEMU without
building any files. The second builds a default set of files and then
exits with an error stating the "--no-build" argument is invalid.

From an implementation point of view, the difference is understandable:
the "qemu" wrapper script understands the "--no-build" argument, while
the "run-zircon" script gets the others. It is, however, rather
confusing from a usability point of view.

Update the tools/devshell/qemu script to parse the full set of command
line arguments, passing the remainder through to the run-zircon script.
This means that both the above command lines have the same behaviour.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509168
Original-Revision: 65bf05e813de07a5035b07ba347c2164cba80841
GitOrigin-RevId: 256700825234a6a38bd6cec0b34ebeffd0ff50be
Change-Id: I7dd464e403a166e153e3131a87db5e6048098379