[roll] Roll fuchsia [blobfs] Separate vmo from mapper in Blob.
Separate out the vmo from the data mapper in the BlobLoader. This is
required because the new pager uses a vmo managed by the PagedVnode base
class and has a separate mapper. This should have no behavior change.

This requires changing the blob loader to separate out the mapping,
buffer, and vmo in various places.

Add FreeVmo() function to Blob in the old pager case for source compat
with the new pager.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511918
Original-Revision: 7b74b903426a2705af41c5c193897303ac091723
GitOrigin-RevId: 5af074be517e77f7a88b9c48969d2ff1b52d9bc7
Change-Id: I3354e1422d3c5adc1a87a6fa4a60a3dcdcee83de