[roll] Roll fuchsia [wlan][policy] add lookup_compatible
This change adds a function to lookup all configs that can be used to
connect to a network with the provided detailed security type (as
opposed to the lookup function that uses a simplified policy security
type). The configs that are compatible are the ones whos security type
can be upgraded to match with the provided security type, ie we upgrade
wpa2 to wpa3 so if we lookup compatible configs for WPA3 we could return
one saved as WPA2.

Original-Bug: 73578
Test: added unit tests
	fx test wlancfg-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512029
Original-Revision: fe4f3f9cf2468336cce4ab5b337d326b318d0285
GitOrigin-RevId: 61213e0feff8d2d6c1c82bbf7930e90225a83916
Change-Id: I961672f38b606dc3dc551f7605e15cd9ee4e7468