[roll] Roll fuchsia [logs] Remove arc-swap dependency from Rust backend.
There's a newer version we're looking at updating to, but it has a lot
more unsafe and will be very high effort to verify. It's only used here,
so we're going to be a bit risk averse and just remove it from the tree.

AFAIK there's nothing that will suffer seriously when going from an
ArcSwap<Severity> to a RwLock<Severity>, and the .read() calls might
even be faster than ArcSwap's fast path when uncontended.

A future change will actually remove the crate, as its still depended on
by an out of date version of signal-hook-registry which will we'll
update past in fxr/512238.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514600
Original-Revision: 3894c033cec0344e653ba7bf63ee4d7682146d79
GitOrigin-RevId: c69458cc58c7682cc8e50318d3e4b8551369d100
Change-Id: I088009ac4b114a4d6ca6e3bca4e300bfdbe46229