[roll] Roll fuchsia [root_presenter] Report focus changes to the keyboard subsystem
The keyboard subsystem needs to know when the focus changes so that
it can dispatch key events correctly.  Root presenter's old code did
not report the focus changes, so this commit is adding a connection
that does just that.

In the process, I discovered that the `FocusChain` spec claims that
`focus_chain` is always set, which is not actually the case: early
in the setup we may end up with unset `focus_chain`, so modified the
protocol spec for `FocusChain` to document this.

Tested: fx test //src/ui/bin/root_presenter/tests
  fx make-integration-patch ...
  fx ... lsc ...
Original-Fixed: 70966
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511640
Original-Revision: cb8b8779ae642fd6d4273b243ba08db5a1a8203b
GitOrigin-RevId: c78231c3a377374bd89049774e55441d5a857575
Change-Id: I1e868dc4ac4bdf68709ba14d0bd80f978f4c587d