[roll] Roll fuchsia [fx] Correct the docs and usage of FUCSHIA_DIR
Correct the documentation for FUCSHIA_DIR to note that the
environmental variable is not used.

Clarify the output message when the Fuchsia root is not found
above the current directory.

Update other copies of the root directory search code for
consistency with the recent deprecation of the FUCSHIA_DIR
environmental variable.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513296
Original-Revision: 319f61507650a955763c008706ff89b225f91ff8
GitOrigin-RevId: 9e8428c5282a215980a38684fb15251cfa1d6222
Change-Id: Iab2fb8bda6893d9c06d8ace0f4772c0ad07371da