[roll] Roll fuchsia [host-target-testing] Support lazily fetching blobs
Some of the OTA test steps do not need all the blobs, like when we pave N-1
onto a device, then validate with sl4f. This introduces a BlobFetchMode,
where if we get a repository with LazilyFetchBlobs, we will only download
blobs when they're requested by the test device. This can half the artifacts
we need to download when getting the old version.

To go even further, we could be even smarter and compare the blobs from
N-1 and N, and only prefetch the blobs that are only in N, but we can address
that optimization down the road.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512311
Original-Revision: 342509143d2a9f7d1c980b08c3d00fd56550dc0b
GitOrigin-RevId: 39c3fb769ce83122a464ca2ea16f78af66b316be
Change-Id: Ia220359872c7db17fcdffcb81ff3e93ae767f71a