[roll] Roll fuchsia [blobfs] Clean up paged BlobLoader path.
Separates out the result of loading a paged vs unpaged blob. The unpaged
path is unchanged.

The paged path is changed to return the PagerInfo and BlobLayout so the
caller (Blob in this case) can subsequently create the vmo and mapping.
This eliminates the need for the vmo creation callback which
significantly simplifies the structure of thise code.

This eliminates the vmo handle duplication in the new (unused) pager
code path which is the source of bugs.

Moves handling of null blobs out of the BlobLoader and into its caller
(the Blob).

There should be no observable behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512862
Original-Revision: 568e8347f3981007b6e17aa8810d3ca18b2d1ecd
GitOrigin-RevId: 98e9d939fa33a55133aed0100b0363101bca791f
Change-Id: I717e27df376a02a650235c30d68c76d113be99d6