[roll] Roll fuchsia [timekeeper] Refactor EstimateUpdated diagnostics event.
In the future we'll have both events from the kalman filter estimate
and from the frequency estimate algorithm so "EstimateUpdate" was not
sufficiently precise. Renamed to "KalmanFilterUpdae".

This also moves the content from a single "offset" field (that
assumed the utc line was parallel to the monotonic line) to a pair of
monotonic/utc which fully specifies a point on the line. Changes to
the frequency will be recorded separately in a different event type
and so don't need to be included.

Original-Bug: 56868
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511041
Original-Revision: bac16821a7480b7cace04317890f0dc8516eed0f
GitOrigin-RevId: 9a16b7e1dbbb5c74474033df3ff7814d3329a278
Change-Id: Idc648f173250e576d377d9d134a8e23a3d2278d0