[roll] Roll fuchsia [fint] Record ninja duration to artifacts
Some dashboards rely on the "build.ninja.fuchsia" step name to measure
the time that ninja takes to build fuchsia. I would like to simplify the
step name to "build.ninja", and ideally remove the dependency on
specific step names, which is fragile, and instead expose the ninja
duration as a build output property.

However, to get an accurate ninja duration we need to exclude `fint
build`'s post-processing steps (which may take up to ~30 seconds) from
the duration, hence the need to record the ninja duration within fint
and expose it to recipes.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514960
Original-Revision: 2e735d0a666a17aaf13273820482f59158484697
GitOrigin-RevId: 4b2472dde0131de066bbb6bf9b21ba1bdfcd6102
Change-Id: I49046db370c627e35e7afd850b960c87127892d1