[roll] Roll fuchsia [blobfs] Improve test setup.
The new pager requires a Vfs which is a more involved setup procedure
than most of the tests were doing. This adds a new test helper class to
manage Blobfs setup and running, and uses it in a numbe of tests.

Remove CreateTest.ValidSuperblock because it was just running the
standard test setup code.

There is no non-test change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515321
Original-Revision: 7b2cc6b8ee4a1ef7592fcc1300f0918037a93787
GitOrigin-RevId: 7b66e979f2780488bd886318b0cd4d944ac5dcb4
Change-Id: I015d00423a14dfdbbb9ff010f6dd56f18ae9a058