[roll] Roll fuchsia [banjo] Add heuristic for generating callbacks
FIDL IR does not contain attributes for external types, so we cannot
know if the `ddk-callback` value is set for the `BanjoProtocol`
attribute. As a result, we can't codegen the appropriate struct name
with the fidlgen backend. As a workaround, we've added a heuristic
whereby the Callback suffix in the protocol type name also signals that
it is a `ddk-callback`.

Tested: fx test src/devices/tools/
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512678
Original-Revision: 70d4952a0588dc5915ed619c8a6508f19165c4e9
GitOrigin-RevId: b13fb1869765520229b1b52b846ac0169ce40764
Change-Id: If1326413bc63ee1b8028691c025d3a76d18d8a76