[roll] Roll fuchsia [fidlc] Allow COMPILE_AND_CONVERT tests for libraries with dependencies
Several new conversion validating macros have been added:
* ASSERT_COMPILED_AND_CONVERT_INTO takes an uninitialized TestLibrary
and populates it with the result of compiling a converted file. This is
useful for converting a library that will be consumed as a dependency of
another library.
* ASSERT_COMPILED_AND_CLONE_INTO is identical to the above, except that
it does not convert the second library (ie, it clones it instead). This
is necessary because we will need two copies of the dependent library:
one to successfully complete the pre-conversion compilation of the
target library, and one to use as an unconverted dependency for its
converted version.
* ASSERT_COMPILED_AND_CONVERT_WITH_DEP allows a library to be converted
with a dependency, generated from one of the two macros above.
* ASSERT_COMPILED_AND_CONVERT_WITH_DEP_INTO is not used directly - it is
only called from inside one of ASSERT_COMPILED_AND_CONVERT_INTO,
ASSERT_COMPILED_AND_CONVERT_WITH_DEP, or ASSERT_COMPILED_AND_CONVERT.

Test: fidldev test fidlc
Original-Bug: 72924
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513299
Original-Revision: 71c59d55a7eb5e1568ad925eba7ba3714eecf8ab
GitOrigin-RevId: 087b1edce5293090a1a07bbed622610c0d7d2982
Change-Id: Ib4fbca092472b972f5c218ea500e2a764fc7a50a