[roll] Roll fuchsia [bt][gap-bredr] Log early initiator pairing failures
In PairingState, log any error code that is in an Authentication
Complete event, instead of emitting only a generic "unexpected event"
error. This case occurs when a peer is not ready to pair or timed out
during user-interactive pairing.

Additionally, should PairingState cause a link to be disconnected, log
the Authentication Complete status code in the "link not found" case as
well.

Document this possibility in the header comments.

Original-Bug: fxbug.dev/74237
Test: fx test //src/connectivity/bluetooth/core/bt-host
In bt-host-gap-tests,
GAP_PairingStateTest.AuthenticationCompleteWithErrorCodeReceivedEarlyFailsPairing
covers the new behavior of reporting the early exit error code.
Otherwise the change is logging-only.
Manually observe the new logging statements are emitted when pairing
fails or does not start.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515612
Original-Revision: 5b70eee6c9676282e9f37f890b252de1addadd4a
GitOrigin-RevId: ca7d3f339dbdad3f5d996ad4881854544a968c05
Change-Id: I37c59ec288f5837332ed1b83411ce629ecd4c6e3