[roll] Roll fuchsia [kernel][lib/arch] Add Write() convenience method to SysReg.
Add a Write() convenience method to the SysRegDerived class, allowing
many instances of:

  arch::FooReg::Modify([](auto& reg) {
    reg.set_a(true).set_b(true);
  });

to be written instead as:

  arch::FooReg::Read().set_a(true).set_b(true).Write();

which is a slight improvement in readability.

Like the existing Modify convenience method, this only works if mocking
support is not required in the code.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513622
Original-Revision: d4118f9472f1f9b6830b9c392a55ef6d6885e3af
GitOrigin-RevId: a75f05827fa6f4b074da454a8285ca7853ac9c83
Change-Id: I1af177d89fdf975cf90797e5912bc10aa6cb3c29