[roll] Roll fuchsia [fidl][llcpp] Do not assume heap allocated ResponseContext
The ResponseContext classes is central to caller-allocated flavors of
async calls - although not yet used in production, it is possible to
allocate the bookkeeping state for an async method through ways
beside the heap (stack alloc, using allocator, etc.), thus performing
an entire async call without any heap allocation.

There is one place in llcpp_message.cc that assumes ResponseContexts are
always allocated on the heap, and tries to delete it on error. This CL
fixes that to properly release ownership by calling |OnError|.

Improved the documentation on the lifecycle and ownership of
ResponseContext.

TEST: added one in integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514113
Original-Revision: 01a97be133cc3b9581d017fa1c4ededec3a43c1f
GitOrigin-RevId: 9493ed436a7ac39ea0cd85f0826760f80ede7924
Change-Id: I8f3d2a5481865afd2966e32905016cce8dbbad97