[roll] Roll fuchsia [bsdsocket_test] Wait for the enqueue of passive connection
This test expects to accept a connection that is reset after it was
enqueued. Deflake the test by ensuring that the connection is reset
only after a successful enqueue. When the listener is using syncookies,
the RST could arrive before the final ACK of the handshake is processed,
causing it to get dropped by the listener.

Original-Fixed: 74478
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516147
Original-Revision: 8c9412259ac657911edac3c65216e46e2cd92d3c
GitOrigin-RevId: 394dbe36f22c68716026f14a0133a68bbcbbb72e
Change-Id: Idedee43c2ff75ac8cff205d37b812d9ec1434295