[roll] Roll fuchsia [fint] Emit ninja targets from `fint build`
I realized I forgot to start setting this field when I added it to the
fint build artifacts proto. Fortunately it's not load-bearing yet, but
will be after fxrev.dev/515584 lands, so this needs to land first.

This is another sign that this function really needs unit tests, but
deleting all the recipe logic that's now duplicated in fint is a
higher-priority task that this fix unblocks. Once that cleanup is done
I'll circle back and add tests here.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516223
Original-Revision: ff617a51e58abf8410ef7134d9980c9c1528ab2f
GitOrigin-RevId: 268dfe100a9df62082846f433d234b2100dc542a
Change-Id: I925345b6040cbe3f381ff7ad7d60fb62e42381a0