[roll] Roll fuchsia [bt][gap-bredr] Handle early pairing failures specifically
In PairingState, log and proxy errors received within Simple Pairing
Complete events that are received "early," such as when a peer does not
want to pair because our IO Capability does not meet its requirements.

This was already handled as an "unexpected event" so the difference is
in logging and the error code returned to the pairing client.

Also document these possibilities in the header.

Original-Bug: fxbug.dev/74237
Test: in bt-host-gap-tests,
GAP_PairingStateTest.SimplePairingCompleteWithErrorCodeReceivedEarlyFailsPairing

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515615
Original-Revision: fd5a6656583a18153139c2072c615bf5b65d75f8
GitOrigin-RevId: 73338d45062334d3008e145f7106088a9a986791
Change-Id: I9f651e4d9d6c5fa16e1a23483a9c595c24c0c70b