[roll] Roll fuchsia [audio_core] Dest gain ramping
When a stream is ramped at the same time that its usage changes, there
is a collision in the Gain object's ramp values, leading to inconsistent
behavior. These should be accounted for separately instead of being
combined into the source_gain side of Gain, while the dest_gain side
sits unused.  One reason dest_gain was not used was that it did not
implement ramping.

This CL adds SetDestGainWithRamp to the Gain object, as well as the
source_mute_ boolean. The two gain components (corresponding to the
object's source and dest) continue to be entirely independent, and
are now both orthogonally rampable.

AudioCore does not yet take advantage of this capability; that will
be in a subsequent CL.

New unittests added to audio_mixer_unittests

fx test -o audio_mixer_unittests
Original-Bug: 74275

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514659
Original-Revision: 0486b4123c5f9ec6f187678aa24ea83bd38d5794
GitOrigin-RevId: 0d1bf5f0885b2ee55d053bd686e076e3dfba3bec
Change-Id: I310c0053343f9ce94a208b34aa31bad2511e5479