[roll] Roll fuchsia [kernel][arm64] Remove code saving the percpu pointer around SMC calls
Previously, the percpu pointer (x15) needed to be saved acround SMC
calls because it was possible that it was trashed in the call. This also
meant we needed to disable preemption which turned out to be a subtle
bug.

After moving the kernel's percpu pointer away from x15 to x20, which is
callee and SMC saved, we can remove the need to save and restore this
register.

Also no longer need to save the shadow call pointer (x18) across SMC calls
because it was always callee saved.

Removed the lockup detector logic in the high level smc call wrapper
logic. Filed bug to track future change to add a new version of that
functionality back.

Original-Bug: 69037
Related bug: 74553

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513645
Original-Revision: 2745ab5f797e85a16c9a8c588ae6e651daebf086
GitOrigin-RevId: 24a6f350812e177b5de55129f0e11e0ac5b8fa61
Change-Id: Ie6a1f745930e6155c002333c39db639297081222