[roll] Roll fuchsia [dhcpd] Fix lease expiration/release inconsistency
When a lease expires, dhcpd needs to ensure that the lease record
reflects that the client no longer holds the address, and that the
address is returned to available address pool. Similarly, when a lease
expires, dhcpd should, per RFC 2131, retain a record of the expired
lease rather than delete it. Previously, we did not consistently ensure
that the changes to the state of the address pool and changes to the
state of a lease record were in sync, and we did not retain lease
records (much) past lease expiration. To address these problems:

- Update CachedConfig to have current and previous fields, both
Option<Ipv4Addr>
- add release_lease() to keep the lease record, the address pool, and
the data store in sync
- replace calls to release_addr() with release_lease(), except in
dispatch_clear_leases(), where we really do want to eliminate lease
records
- reimplement release_expired_leases() in terms of release_lease()
- update get_addr(), now get_offered(), to correctly follow RFC2131 now
that CachedConfig has a previous address field; remove the call to
release_addr() since it was a hack around store_client_config()
- update store_client_config() to allocate_addr() only if the address
needs to be allocated
- update tests to reflect new behavior
- misc. cleanup

Tested: fx test dhcp-tests && fx test netstack_dhcp_integration_test
Original-Fixed: 74545, 74546
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516120
Original-Revision: 71606963fa8dbb2b843e5394715bab49e9330186
GitOrigin-RevId: 2b8c1de0d3fbaa4b1aaddf1eeb360143f9e1650b
Change-Id: Ib0e15981d95defdc2ce0717cdf69440ecc84427f