[roll] Roll fuchsia [kernel][arm64] slight optimization for hand written assembly in context switch
As a general optimization, if pushing or popping a bunch of words
on the stack or other structure, it's better to adjust the base first
and then use a series of load/stores with an immediate offset.

Adjust the core context switch routine to do this.

References:
cortex-a55 optimization guide, for simpler, dual issue cores.
cortex-a72 and above optimize guides, for out of order cores.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514042
Original-Revision: cd1e967c3c9fcf5beda648958c488b582409d5b3
GitOrigin-RevId: 993ddaad7598a9b806c478b3d8c561969f2ae99c
Change-Id: I8a817e49dec2d5f8f09defc4cb6c559a3a0a8618