[roll] Roll fuchsia [fidlgen] Drop shared NameContext
This is a retreat and refactor from the previous CL
Ia296aa281d1377d1cebfb291d590427ccd6f541e which moved all of the name
context logic into the fidlgen package. While trying to use this code
with the C++ bindings it became clear that it makes sense to share the
code to initialize and track the sets of reserved names, each binding
has a different sense of what other data to associate with a name
context.

In this CL fidlgen.ReservedNames just holds the set of reserved names
and has an IsReserved func to determine if a name has been marked as
reserved and a ReserveNames method to add a slice of names to the
reserved list. fidlgen_dart.reservedNames is a simple wrapper that adds
a changeIfReserved func that applies the Dart binding's renaming rules
if needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515361
Original-Revision: 74ef03d3c213bee8d9c905f1b687fc9eab511a25
GitOrigin-RevId: 910761e782b52b47e5b3ebaeb6c21b33a81d5464
Change-Id: I2c22313d115a25dd5759d84fb2bf038d7e54402a