[roll] Roll fuchsia [stress_tests] Swallow all errors in storage stress tests
When an actor in a storage stress test gets an unexpected
status code, assume that it is because of an intentional crash.

The environment will verify that an intentional crash occurred
during reset. If this is not the case, a panic will occur.

This should help with the whack-a-mole situation we're having
where we have to continuously allowlist more status codes
that show up during an intentional crash.

Multiply: blobfs-short: 100
Multiply: minfs-short: 100
Multiply: fvm-short: 100
Test: fx test blobfs-crash
      fx test minfs-crash
      fx test fvm-crash
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516280
Original-Revision: 4dc95dbcdb2e6011c1a9ab8bc17f9982a7badb74
GitOrigin-RevId: ec71c80ef3bd7d99f11c41a6f86a1867f31c3177
Change-Id: I140dd42432464456caac1a02302fa838e147ff15