[roll] Roll fuchsia [rust] Use backtrace-sys mirror
Update and use the backtrace-sys tiny mirror instead of duplicating it
with a vendored copy. This crate is a wrapper around libbacktrace for
use with the backtrace crate. The library isn't intended to be used with
Fuchsia anyway:
https://github.com/rust-lang/backtrace-rs/commit/6bc5be152e8b8b31e8c3236326c60f5d8485c942

This removes the copy of Tom Sawyer, which isn't in line with our policy
on inclusive language.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511725
Original-Revision: 38f540f8435ef503791553f1ea512af9be1d2f48
GitOrigin-RevId: 5ea7db3c5b4b271fbf277bb8bc42f8e49bd06ba5
Change-Id: I4f5d848665998ca7d5ee6020a84bea2132e11b50