[roll] Roll fuchsia [fdio] Add fdio_ns_is_bound
Add a function for querying whether a given path is bound to a remote
directory in a local namespace. This function was requested by a
customer to check whether a directory capability was provided in a way
that is guaranteed to be non-blocking.

Original-Bug: 74495
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515730
Original-Revision: 63eda70d3471f970d9b8f9681bb885877591444c
GitOrigin-RevId: 23f30169ca2d8ca2a27d369374e9fa439df600d8
Change-Id: Ied305e87445fe69593b5b8f837b90c6eb92f68db