[roll] Roll fuchsia [bringup] Add virtcon2
This is the initial implementation of a Carnelian based
virtual console. See README.md for more details on how
it is different from legacy Virtcon and the transition
plan.

This initial version is not meant to be a full replacement
for legacy Virtcon. It demonstrates efficient use of
vector graphics, and provides a dumb implementation of the
session manager service in order to demonstrate how
functionality in the Terminal app can be extracted into
libraries for code reuse.

Test: fx test virtual_console_tests
Original-Bug: 74628
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511880
Original-Revision: c1c69ec8d31bb969e16f49ff5d776cbbf052617a
GitOrigin-RevId: 3055e15d5318220f2caf315818754c0bf2e35775
Change-Id: I236b2c92a2b1eae4d940bd1c7ce221d610f5fe33