[roll] Roll fuchsia [kernel][lib/arch] Make arch::InitializeBootCpuid C++-friendly
This function was only originally meant to be called in assembly. Now
that we have a C++ use-case for it, we update it to ensure that ebx/rbx
- caller-saved - are indeed saved (previously they were being thrashed).
While we're at it, we add CFI metadata. Note: the function is still
buggy on repeated calls after this change; a follow-up will address
that.

Original-Bug: 74527
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516061
Original-Revision: f59f45c4328070cd2c6c8efbeda28ff2584594be
GitOrigin-RevId: 88039ec6b7be143f45d69ba10c383c526dfa944d
Change-Id: I167221a4a8e3c875d5e9f8239cf51a0f77666708