[roll] Roll fuchsia [starnix] Update example hello world component
This new component is built as part of the Fuchsia build, so we can
remove a manual step for testing starnix. It also defines a test
component using the same binary, that is run using the starnix test
runner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516380
Original-Revision: b6f2348708b076f00e2ad7aebbeaac47a17fe3e0
GitOrigin-RevId: 482281b0f5737e4078c9604831782b8fe28f7b85
Change-Id: I9c9585c98e27ff9eb2a1fc2624cad44cac46b4c7