[roll] Roll fuchsia [blobfs] Stop storing the inode in the blob
Storing the inode in the blob as we were was prone to error because
after writing a blob, it isn't in sync with what has been written; only
some of the fields can be trusted. It's also a bit wasteful on memory.

Original-Fixed: 74799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517440
Original-Revision: 1d810c28bb1c78593f0bce1a8f5c370eeb318c3e
GitOrigin-RevId: 183d1c33546ae5dd3946ac1cab072be2c778235b
Change-Id: I683a9744f2c002a0c54bd3bed895e2e8a97b6c43