[roll] Roll fuchsia [benchmarks] Remove step that rounds values to ints when summarising data
The rounding was done on the assumption that we don't care about
sub-nanosecond variation in test results, but this has turned out to
be not quite right for results with low variation.

These result values are themselves averages, so the fractions can be
meaningful, and it's useful to be able to visualise the fractional
values in graphs (such as in Chromeperf) rather than seeing the metric
distributed across a small number of integer values.  Also, rounding
to int potentially makes the value less normally distributed, which
might interfere with applying t-tests.

There may still be some benefit in reducing precision to save space,
especially when storing results from many builds together, but that
reduction of precision can be done in a different layer.

Original-Bug: 68471
Original-Bug: 74469
Test: fx test host_x64/summarize_perf_results_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517682
Original-Revision: 7828a32b62c925534d2a06dffa89b209a4dacafb
GitOrigin-RevId: 8baf38cf0d31095429c5149c2a72a6c433cf1d8c
Change-Id: Ia67f9b0248d64a9d2d770f9f606b7c2a90d14823