[roll] Roll fuchsia [docs] Fix "Formerly known as FTP-NNN" notices
This CL edits RFC-0026 and RFC-0036 to add "Formerly known as FTP-NNN"
notices, which were missing. It also moves these notices to the top of
the page, below "## Rejection rationale" instead of below "## Summary",
for all rejected RFCs. It used to be inconsistent, with 7 rejected RFCs
using the former style and 4 RFCs (0031, 0034, 0038, and 0044) using the
latter. Now all 11 use the former.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517141
Original-Revision: b6687adc2e1fcb99a96010303b63af1336dd0474
GitOrigin-RevId: edb7c7ddaba6a93f16f805134fb75990c43eaf00
Change-Id: Ia2f9bfb0f34d72eff1266bb39c1cd2e20d950f9c