[roll] Roll fuchsia [kernel][arm64][event stream] enable by default
A simple inversion of the default of the command line option to
enable the event stream timer feature from false to true.

There's strong evidence that leaving the option on all of the time is
safe and generally a good idea.

Original-Bug: 74457

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515403
Original-Revision: d6e841526f852907720841e815c7e80256b279b6
GitOrigin-RevId: 147cbdb0b4312b0ea03815faa47e1bd74e24b86c
Change-Id: Iabd15c662be48fa65acc09a5d27853ebba4742ad