[roll] Roll fuchsia [kernel] Migrate kernel.entropy-test.*
Added a constant to defined the default max, and wrapped the length on a
function that returns the macro value when defined for the test.

`quality_test.cc` was bit rotten, had to update it so it could build.

Created a bug to either add a build only bot or remove it:
fxb/74705

Original-Bug: 53594
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516781
Original-Revision: b6e10bed23898ca8e03f30a54024fb717b075c42
GitOrigin-RevId: 8964f7c5beb72f0212c4fb112dcbf2f455ddd1dc
Change-Id: I552a5822e5ecf0605d738050ceda350a0d5adac7