[roll] Roll infra/recipes Delete unnecessary trailing commas
When we update to v20.8b1, Black will start exploding collection
literals with trailing commas onto multiple lines. This is useful for
collections with multiple elements that don't fit on a single line, but
it's awkward for very short collections that can fit on a single line.

So before updating Black, remove all such commas to prevent splitting
these literals onto multiple lines.

I did this with a series of regex search-and-replace commands in VS
Code:
- ",]" -> "]"
- ",}" -> "}"
- ",([^(]*),\)" -> ",$1)"

The parenthesis case is tricky because the trailing comma is actually
semantically significant, indicating that parenthesis around a single
value should be interpreted as a one-tuple rather than as a no-op. So I
had to make sure to only remove trailing commas from multi-element
tuples.

I looked over the files after making the change, and the only manual
fix required as to add back a trailing comma in a JSON string in a
jsonutil module test that specifically tests for trailing comma
handling.

Original-Bug: 74877
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/517703
Original-Revision: 4db253c58d6e7c0f7ed658692481d40516a5b795
GitOrigin-RevId: 04af02fb50dd140a82c3ad5a25f704481196b2bc
Change-Id: I8f45760843ca948c0cb2085be8f86555627b687f