[roll] Roll fuchsia [fxfs] Implement unlink
For now, files are always moved to the graveyard. Eventually we can
optimize this for the case where the file has no open references by
immediately deleting the file, but that requires appropriate locking on
the file node which isn't implemented yet.

For now, we don't ever purge files which are placed into the graveyard,
so they'll continue to take up space until we implement that.

Also fixed a small TODO in SimpleAllocator (to not assume an infinite
device), since this was tripped over when writing a test.

Tested: unit.
Multiply: fxfs-tests: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514621
Original-Revision: 80063857936c187a5080d89fa86a73bb2d546942
GitOrigin-RevId: a5b9eac2963c7daff86ca08664ec81fcf0e0fecd
Change-Id: I4ebd466d61bf47bd9f04c6e550da2c0bab21c178