[roll] Roll fuchsia [blobfs] Use shared reads.
Utilizes shared locks for reads which allows multiple read operations
to execute at the same time. This has no effect on the current code
because there is only one fidl thread, and the current (old) pager
doesn't lock with the blob at all. But in the new pager this will allow
potentially more than one pager and/or fidl thread.

The shared locking is also necessary to allow the fidl-based read
operations to be backed by a page request, since this re-enters the
lock.

The paged/unpaged variants of the BlobfsCheckerTest are removed. The
flag was no longer hooked up and these tests were the same.

The BlobCkeckerTests are updated to use the new BlobfsTestSetup class
which does additional work required for the new pager. These tests now
pass in the new pager.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517007
Original-Revision: 2439e23f2a1afd4391fa6bf0adb83cd2bfdea983
GitOrigin-RevId: fb18729ebd5e1d19681fbe1ab5aa3b02eeb544d2
Change-Id: I2607203af6e0005254e5a6bff903c20733133ed0