[roll] Roll fuchsia [fake-clock] Use async task and cancellation
The future attached to an fasync::Task stops being polled when the
handle is dropped, effectively cancelling its execution. This is
guaranteed to stop execution in fake clock as only a single thread is
present, and has the advantage that the code cancelling the task does
not need to be async.

Testing: existing tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517903
Original-Revision: e83c912d0d166982217f8f3026ed848c5407e49c
GitOrigin-RevId: 7e0d88fa46d63cdb634b63a8bbacbec3daf5901a
Change-Id: Ibe0c6fd778930f1734f67535c7d39502c3e2b367