[roll] Roll fuchsia [hypervisor] Remove dependency on Zircon utc_offset.
Zircon no longer maintains a UTC clock and, per RFC-0008, the syscalls
and internal state associated with the old clock will soon be removed
entirely which means we need an alternative solution for hypervisor.

I synced with Abdulla offline and he's in support of removing the usage
of utc_offset given guest OSs have other options for setting their UTC
and noting we could find another way of setting UTC from the userspace
clock in the future if necessary.

This CL is very much a barebones implementation of the removal; none of
the structs are changed, the boot time is fixed at zero and the offset
is just set to monotonic time. I'm not familiar with hypervisor so very
willing to take alternative suggestions if you'd like anything done
different.

Original-Bug: 74471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516330
Original-Revision: 81d891c03f2419a7f9ee9bc3af70647519f8c311
GitOrigin-RevId: 0517dea7fc46f53ca3ca8e5f99b17bd2c0b8ff84
Change-Id: I3873b0f036c3c7708a461dc02d209c9736f77780