[roll] Roll fuchsia [fshost] remove registry
Nothing was using the registry for anything, and we have no plans to use
it in the future. It was originally meant as a new way of routing
filesystem apis. This role is now taken by component_manager.

Original-Fixed: 75028
Test: fx test fshost-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/510507
Original-Revision: f8360ddc2c9c87a4317f6050a34238da05c1b3d1
GitOrigin-RevId: 11bfa22d66b9b3d92a1efbfb36065c60b80c781e
Change-Id: I013d3f0a6bdd643a1a865cb16ecae2ad59de86f0