[roll] Roll fuchsia [kernel][lib/arch] Make arch::InitializeBootCpuid idempotent
We have a use-case for repopulating our CPUID cache and,
absent-mindedly, a second call to arch::InitializeBootCpuid was
introduced to accomplish this. However, in the function's current form,
it just overwrites the cache with nonsense. In this change, we update
the layout of arch::CpuidIo to store the associated leaf and subleaf
and update the population logic to allow for the function to be called
multiple times.

Original-Bug: 74527
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515801
Original-Revision: 750b8c8b188fed024869b8d3259cb8a4ddacc168
GitOrigin-RevId: 646b0aaacbe3ae709fbf4a5e6fa1806e0805509d
Change-Id: I05d85d3b149dfd4ac9fff5ee7f0b257a73c287ce