[roll] Roll fuchsia [kernel][lib/arch] Introduce arch::ApplyX86ErrataWorkarounds()
This change introduces a consolidated
arch::ApplyX86ErrataWorkarounds(), which also refactors the existing
logic specific to AMD family 0x17, models 0x00-0x0f. The intent is that
this function will be extended for other AMD and Intel processors.

We removed the workaround for erratum 1090, as that was sourced from an
NDA-guarded Revision Guide - meaning we were unable to document what it
was doing and why - and as it did not seem critical to mitigate neither
(Linux nor FreeBSD reference it).

Original-Bug: 61093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515360
Original-Revision: 6bfc22e01a7cd4d1a147d58cead9dddd70a623c4
GitOrigin-RevId: 90ec347566245a71776d978a77471866da2a3656
Change-Id: I76b9368c4bf4370d2122d42ea067c62612e1a00f