[roll] Roll fuchsia [artifactory] Upload objs_to_refresh_ttl.txt and set CustomTime on uploads.
This will list the blobs and debug binaries that we skipped uploading
because they already existed in the GCS bucket. A post-processor can
then use this list to determine which objects to update timestamps for
so that they don't expire.
This also sets the CustomTime attribute on all uploads so that
we can change our buckets to expire objects based on their CustomTimes
instead of their creation date.

Original-Bug: 66462
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516680
Original-Revision: ba6357531edb5f3395423e02f30fb06742ad2a54
GitOrigin-RevId: 725a3d7f65b6cf76452acb77e4a6a36041a05644
Change-Id: I20e20c90b55051983e1e80c17a424bf9440f7c94