[roll] Roll fuchsia [input] Fixes some issues related to key ordering
The whole story is a bit involved, please see fxbug.dev/74271 for
a run-down of issues.

The short of it is that the current delivery discipline around
simultaneous events may cause a modifier key to be applied to a key
that preceded it.  It is a side effect of a few design decisions at
different points in the input pipeline and the test fixtures, each
of which quite fine on its own, but when strung together result in
an unwanted behavior.

This change cleans up a few issues mentioned in the bug, and may
incidentally fix some.

This is a prerequisite change to fxr/513823.

Tested: fx ... lsc ...
Multiply: experiences_ermine_session_shell_e2e_test
Original-Bug: 74271
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514048
Original-Revision: bb64b2a18d00dd0bb59760ae1220806319f763a3
GitOrigin-RevId: 10cc2af375bf7a008078fb20a194333b412d4b7c
Change-Id: I489130cfe505f001f062d9b239ad6329aa3e135d