[roll] Roll fuchsia [overnet] Flake fix
The connection from the process running `onet list-links` to `ascendd`
may not be formed nor be in the process of forming when `onet
list-links` polls itself to get a list of links.

This caused a flake in a test that was measuring how many links were
reported.

Since we know that there must be at least two nodes on a mesh for `onet
list-links` to work (the other being ascendd) we also know that there
must be at least one link per process and so it becomes an easy thing to
just wait for things to settle out before proceeding.

BUG: 73091
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518746
Original-Revision: a95a270163b7f8af2974553fc2a1a6c07459e502
GitOrigin-RevId: 21dfbaa3de19c8303e334abef46554f5afe76552
Change-Id: I7d2a084c929fade05aef6f0673f21cbee66f5aa0