[roll] Roll fuchsia [ffx][repository] Bring repository structure plugin in line with rest of tree
There's a convention for subcommand structure in ffx that gives us the
help command automatically as well as a few other niceties. This ports
the framework of the repository plugin to it.

We also add a stub for the list command in addition to the one for the
serve command.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518911
Original-Revision: 124e685e5ba067fe5416a96f2569146822df9d06
GitOrigin-RevId: 7d41825adc57340590f5cedc30b858f5ecae37e2
Change-Id: Id04b3d0e925f02254c8a779b46c7fc6431025394