[roll] Roll fuchsia [fidl][rust] Explicitly test send-side and receive side handle rights checks
Previous handle rights tests in rust only tested end-to-end handle
rights checking. This only demonstrated that the rights checking was
happening *somewhere*, but the requirement is to do checking on both the
send and receive side.

This CL adds separate testing of send side and receive side checking.
Because send-side is not implemented yet in rust, the tests are disabled
and currently do not pass when enabled. The technique for isolating send
side and receive side rights checks is to replace the ordinals during
transit over channels so that one protocol method is used on send side
and a different one on receive side - one with stricter rights checks
than the other.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517802
Original-Revision: 5aa85f44c03ee121928d169f6e483fda8e8a37e2
GitOrigin-RevId: e3d304d03c31145ad24ac53d56d1a5ea04e9354a
Change-Id: I87fc0dd95bdd49146c745b33cf536f6f7c2eaca6