[roll] Roll fuchsia [fidl][llcpp] Dispatcher runtime will not perform decoding
FIDL messages on the read path don't have a standard pattern of
ingestion and decoding:

- In conformance tests [1] and some direct-object-decoding scenarios, a
`fidl::DecodedMessage<T>` is supplied with bytes and handle info, and
decodes the message as part of construction.

- However, message dispatcher in production calls [2] `fidl_decode_etc`
directly on a buffer. This may increase performance but at the cost of
safety:

  - Because the buffer is just a uint8_t*, it can be hard at times to
  determine whether a buffer is in encoded or decoded form. This is
  compounded by the optimization of factoring decoding work into the
  binding runtime [3] for client response and server request decoding,
  but not doing it [4] for event decoding, so the message was already
  decoded when it gets to generated code, except for events.

  - There is no RAII type as the last safety net for closing handles.
  Instead we rely on carefully inserted [5] `_CloseHandles` calls at
  right places in the control flow. This led to handle leakage [6].

The unified bindings add more dependency and complexity into the
LLCPP dispatchers. For example, whereas the LLCPP dispatcher previously
always closes unknown handles, now it needs to support closing unknown
handles when the messaging layer bound at runtime was speaking wire
types, but preserving unknown handles when the bound interface was
speaking natural types. This late-binding behavior can be tricky to
support in the dispatcher runtime.

Based on the above, we would like to move the decoding logic to the
generated messaging layer, such that the message passed to the generated
messaging layer is always in encoded form, for simplicity.
This allows the unified bindings to inject its own decoding logic which
does not close unknown handles.

This CL uses the C `fidl_incoming_msg_t` type to represent an incoming
message in encoded form.

The next CLs will introduce a public `fidl::IncomingMessage` type to
represent this incoming message in encoded form, while implementing
handle ownership via RAII.

[1]: https://cs.opensource.google/fuchsia/fuchsia/+/c43485f9ffe017776d351813082d888c1ee1b07d:src/lib/fidl/llcpp/tests/conformance/conformance_utils.h;l=137
[2]: https://cs.opensource.google/fuchsia/fuchsia/+/c43485f9ffe017776d351813082d888c1ee1b07d:tools/fidl/fidlgen_llcpp/codegen/fragment_client.tmpl.go;l=115
[3]: https://cs.opensource.google/fuchsia/fuchsia/+/c43485f9ffe017776d351813082d888c1ee1b07d:zircon/system/ulib/fidl/llcpp_server.cc;l=17
[4]: https://cs.opensource.google/fuchsia/fuchsia/+/c43485f9ffe017776d351813082d888c1ee1b07d:tools/fidl/fidlgen_llcpp/codegen/fragment_client.tmpl.go;l=115
[5]: https://cs.opensource.google/fuchsia/fuchsia/+/c43485f9ffe017776d351813082d888c1ee1b07d:tools/fidl/fidlgen_llcpp/codegen/fragment_client_async_methods.tmpl.go;l=43
[6]: https://fxbug.dev/74790

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516023
Original-Revision: a84c99efefd0d155f9cf88f039f4d4f68ca3a7b1
GitOrigin-RevId: dfab71aacc4f7cd7d9b3efd82f1a2bb8d9a1b818
Change-Id: Icda092790ed1902e27ab3f9b82ea59a06b36e024