[roll] Roll infra/recipes [snap_branch] Add option to no-op on mismatched milestones
When a branch is created, say F3, we have a MILESTONE roller to detect
the new branch and bump the value from 3 to 4 at tip-of-tree.

At the same time, an automated snapper may continue to find an LKG which
is behind the MILESTONE roll, and create a 3.B.C.D version. This is not
desirable because we do not want to create any more 3.B.C.D versions
off tip-of-tree LKGs after branching F3.

The option `must_match_milestone` checks that the resolved release
version's milestone matches the milestone at HEAD, or otherwise no-ops
with SUCCESS status.

Essentially, when this option is on, automated tip-of-tree snappers will
wait to snap until the CI LKG has caught up the new MILESTONE number.

Original-Bug: 75095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/519420
Original-Revision: 9b87f93007900e945e0c8d5cd4a8778c432d3d32
GitOrigin-RevId: 69dc25e82b46ecb5bdc58ff4dc1cc0105bcd4e39
Change-Id: I8ce711410f31dd205ad6322c150e46d2fedcb8d5