[roll] Roll fuchsia [ime][tests] Puts a band-aid on a test flake.
Adds an... ehm... sleep in the code path for the IME service tests, to
try to reduce the incidence of test flakes due to an inherent race
condition in `ImeService.GetInputMethodEditor`.

The issue is that the above method is a fire-and-forget method, which
means the remote end could in theory take infinitely long to set up its
end of the connection.

If the test fixture starts communicating with the remote end too soon,
it will lead to `PEER_CLOSED` errors, and skipped events, which in turn
will trip up the test assertions.  If the test environment happens to be
just slow enough, the tests could end up hanging altogether.

While it would be possible to instrument the code out of band to add
synchronization, this is not something we'd like to do given
that we are aiming to replace this protocol with a new implementation
that doesn't have this issue.  I think this used not to be a problem in
the past because there were synchronized methods that would ensure the
race doesn't happen.  But as we factored those methods out of
`ImeService`, we can no longer rely on that effect anymore.

For this reason, a `sleep` may be an economical way to proceed.  Not
entirely sure I won't have to eat my words very soon.

Multiply: ime_service_integration_test
Original-Fixed: 73945
Original-Bug: 75030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518906
Original-Revision: 9f9a963af5c02190417282c5b9f57e44ba8e56e9
GitOrigin-RevId: de5685bbc87c3e42f5df11406c7018b473d0fbfe
Change-Id: I14015c9a0d2d050549eeb274817584e79fe1102e