[roll] Roll fuchsia [bt][hfp] Send custom Battery Levels through test API.
Use the HFP test FIDL API to send battery level values to the HFP
component. This simulates the battery level changing on the Fuchsia
device. It will cause the HFP component to send a battery level status
indicator to the peer if the status indicator is activated.

The previous version of hfp::call_manager_init_and_peer_handling
incorrectly expected the the peer task to make a request for network
information. That's beyond the intended scope of the test which was
meant to test only that the hfp component was handling and routing
events in its main loop. It lead to failures when the order in which
the select! loop handled ready futures. Now the test is robust to
any ordering.

Test: SL4F test commands succeed.
Test: PTS receives test command.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515335
Original-Revision: 12aa8e3e367a9e2581a7a57597bedc5c5e898517
GitOrigin-RevId: 481047752382ca7eeaa1fe51b7de4a771a5445d2
Change-Id: I70acc813212bb68aaa6653dda23baa60d47675f0