[roll] Roll fuchsia [fxfs] Improve node lifecycle management
Removes a (very) coarse lock on the node cache, which was previously
held for the duration of disk loads on cache misses.

Implement NodeCache::insert_or_reserve() which returns a placeholder to
the first caller, and blocks all future callers until the placeholder is
filled (or the placeholder is dropped). This is used in
FxVolume::get_or_load_node to block all readers until the node is
loaded.

Multiply: fxfs-tests: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516920
Original-Revision: 7380619da58b6fee3274445d530956a68837aae0
GitOrigin-RevId: 476bc8cf81071405bc33843f67fa6adf9bbede96
Change-Id: Iade03cd60b7b2e581fcd1219802d46512422a1f0