[roll] Roll fuchsia [storage] Change io2's Unlink method.
The unlink method needs some options so that clients can assert the
target is a directory.

We also change the method so that it specifies a name rather than a
path.  All our current implementations don't support path traversal
server-side, so this shouldn't be an issue.  If we need to support
traversal later (perhaps for performance reasons), we can change it back
and since it would be implemented within VFS, it shouldn't be too
problematic to change it.

Original-Bug: 74923
Original-Bug: 74911
Original-Bug: 75137
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518063
Original-Revision: 6fd7b2c6d044779d18f790e961e0ca03efc02df8
GitOrigin-RevId: dda02795200c663dcabc90c07597e096a28a9bd5
Change-Id: I630ad8e893dc6f0282faee3022d9a5c3bb1d3217