[roll] Roll fuchsia [kernel][crashlog] Use FILE objects during crashlog recovery
When creating is filling out the VMO which will hold a recovered
crashlog, use a kernel FILE object to render to instead of using a
callback and a context pointer.

This change eliminates the need for a moderately sized stack buffer
when creating the crashlog header, and also will make it a bit easier
easier to start to incorporate other sources of data in the rendered
crashlog in future CLs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506929
Original-Revision: 55f61a5f579fb0ce0208a4c2d3a46fb1f7284829
GitOrigin-RevId: e6df79fa3d005da1a19f530812e04aeb10f514c9
Change-Id: I0c26d5c24b0192c50c4eb7407f4adffd38bce8e5