[roll] Roll fuchsia [fidl][summarize] Harmonize fidl_api_summarize with RFC-076
In the course of reviewing what now is RFC-076, it was decided to add
the values of constants, enums, bits and struct elements with default
value to the API summary.

This change modifies the program code to produce matching summaries.

Tests are modified to match; tests are added to verify the new use cases
for value changes.

Original-Bug: 72179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519981
Original-Revision: 17171847c48ee43c9973e399569cb3eb5989b707
GitOrigin-RevId: fa87fcd6e218e523c2d4dc9cf7d3a420f60ff73f
Change-Id: I5e601de1226eae80dcf68bce1795127a6e34629c