[roll] Roll fuchsia [virtualization][guest] Send errors of guest binary to stderr.
Currently, errors of the "guest" CLI binary are often sent to logging,
instead of the CLI. This may lead to user confusion, due to silent
failure of the command.

Modify guest to output to stderr instead of logs, given that the guest
binary is a CLI tool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519867
Original-Revision: 92b6b7370f8fd6229200485d34ba5acf7d4b6443
GitOrigin-RevId: da08b46baf9ffbdfed3ef59dd280204a1f05a111
Change-Id: I3ec16256616b046f8e4d4b74d126bc416dbba92f