[roll] Roll fuchsia [run-test-suite] Define interface for persisting scoped artifacts
Defines and leaves a space for an interface for persisting artifacts
scoped to a test run, suite, or test case to a directory.
Writing results "live" to stdout is not included as part of this
interface.

Up next: Add a preliminary implementation for writing to directories.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513992
Original-Revision: 1c5a190bebb9df91f1cc01923056140b53eec5dc
GitOrigin-RevId: dd6a74b429d83e6199379c7ae8a014b6abda2580
Change-Id: Ie64154da5524219fc04ca29f4a3b5f7f78282742