[roll] Roll fuchsia [scenic][ui-input] Add inspect to FocusManager
Adds an inspect node to the FocusManager and starts tracking the
focus chain.

Also performs a little bit of cleanup of the focus chain assignment
logic.

Won't be visible in inspect until the FocusManager is hooked up to the
rest of Scenic.

Original-Bug: 75018
Test: fx test focus_unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519664
Original-Revision: 0383875aa2d50b1824ccb011944715bb3823ccc8
GitOrigin-RevId: 3cfd86459b4e0e90168adb40db6690785d55c13a
Change-Id: Ib64dac4a3a66ca86731ca5f86060e87382428cce