[roll] Roll fuchsia [volume_image]: Fix Off By One.
The problem was that data_slice slice(the first slice that is not backed
up by data and is not filled) was not falling into the right branch,
since the comparison ignored it (> data_slice_count) instead of >=.

Then on the next iteration, it will add the range including data_slice.
This was tricky since most of the time we have zero filled, and even
sometimes the order of the partitions made it so this wouldnt be
detected.

In the end, we can reproduce by having a mapping with no fillings, and 3
slices, with one backed by data.

Added an assert that would blow up without this fix.

Also, this tickled a bug in MtdWriterTest, since it now has empty slices
not incorrectly written as before, which has a different behavior on
nand and a file. A truncated file will assume that an unwritten offset
as this jump from the current end of file to further ahead is zero. On
the other hand, on a device, such as raw nand this is tread as unwritten
spaces. Because we are using a FTL, this ends up meaning that a range
is unmapped, which means that the contents are the same of an 'erased'
page, that is 0xFF.

Test: storage-volume-image-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520060
Original-Revision: 8be81961e17925cb9e684d9854d87ee96723c3f7
GitOrigin-RevId: a0ca9108e5b7a39318208d59ef2d56a5ce671d85
Change-Id: Idc111dbb04f32f14354161de3a9599dccff4cdbe