[roll] Roll fuchsia [input] Fix comments regarding "machine uptime"
The `event_time` field corresponds to the monotonic time as determined
by the zx_clock_get_monotonic syscall.

This is not the same as machine uptime. It does not always reset on
reboot and does not adjust during sleep, and thus should not be used
as a reliable source of uptime.

https://fuchsia.dev/fuchsia-src/reference/syscalls/clock_get_monotonic
has more information on this.

Test: N/A
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519480
Original-Revision: 9ff52e332e1edf48cc8e4a30e3359598801b8c4c
GitOrigin-RevId: fb1a1b529efc27230cb9df43088f9391d88d249b
Change-Id: I4ff1ec5fdb7f0dd0c76c07443db54f3209a934e2