[roll] Roll fuchsia [graphics] Support loading ICDs from packages.
The loader will connect to components to get information about their
ICDs, including a new manifest.json file. These ICD files can be used as
a fallback if the ICD shared library doesn't currently exist in
/system/lib. This will allow a soft transition from ICDs in the system
image to ICDs in a file.

The loader also implements an ICD runner, which is a simple way for a
driver package to serve its package contents. The interface with the ICD
runner is described in README.md.

Original-Bug: 13078

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515640
Original-Revision: 31c5dae709230ed1135943c36218ef6fdc8be92f
GitOrigin-RevId: 5246af98b2598622b90a407e0d64c4334fa4f4c3
Change-Id: I9a90b9447488866745fcd6667c3695d3422e9b5a