[roll] Roll fuchsia [fidlc] Add element_memcpy_compatiblilty to vector coding table
element_memcpy_compatibility indicates if the data in the vector body
is possible to memcpy to the channel message, i.e. it doesn't require
any mutations. This new property element_memcpy_compatibility is
calculated off available type shape data

This will be used in the iovec encoder to detect when messages can
be directly pointed to by an iovec that is sent to the kernel. If the
encoder needs to mutate data, then performing these mutations on an
object that is not copied would break the invariant that the input
object isn't mutated (other than to move handles).

Original-Bug: 66977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519380
Original-Revision: 93833666dbe7899f48c60f8521b54745599cdb5d
GitOrigin-RevId: f9290646cc60b8df4fb2deb36396bd8f4ef67578
Change-Id: I8d68a87c729d9899010159537066a150b2f8cc7c