[roll] Roll fuchsia [optee] Move OpteeClients to separate async loop
The optee driver was creating additional device instances for each
fuchsia.tee.Application and fuchsia.tee.DeviceInfo binding. This CL
creates a separate async::Loop with a single thread for servicing these
bindings. The existing deadline profile is applied to the new thread, as
it will be servicing trusted application requests. The OpteeDeviceInfo
class was eliminated in favor of serving the fuchsia.tee.DeviceInfo
protocol directly from the OpteeController class, as there is no server
state for that protocol.

Additionally, a test was added to ensure that any open TA sessions are
closed at the destruction of an OpteeClient, so that all OP-TEE
session resources are appropriately released.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513167
Original-Revision: 992a989ce2c4e591450439def426563f97157745
GitOrigin-RevId: 0ecd535ec5c778d670b6a4f65414d47673e7743d
Change-Id: I6203e3b63296a74e1c7a23b06a900a0bdcf0dcd6