[roll] Roll fuchsia [ffx][repository] Implement ffx repository serve
This is the initial implementation of `ffx repository serve`, which
allows the repository plugin to serve a package repository from a
directory. This is a short term command to ease the migration from `pm
serve` over to an `ffx` based command.

Original-Fixed: 74380
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517743
Original-Revision: a1f2058f2b9f17109283ffbda60103302f2d7418
GitOrigin-RevId: 84632d2653f92e479ae563fb45a3eae9a449521b
Change-Id: If0f2d33df20b251cf9a1a4b0f47fa14ccd5d342a