[roll] Roll fuchsia [fidlc] Change type used to store resolved handle rights.
In FIDL, handle rights and size bounds are similar: they are
both Constants that fall back to a default when unspecified
(same rights, and max size respectively).

This CL updates the compiler representation of handle rights
through the compilation process to be more consistent with that
of size bounds: we add a specific type alias on the ConstantValue
type used to represent handle rights, and use a pointer to this
to refer to a "resolved" handle rights value instead of storing
the top level Constant.

Besides adding consistency, by storing the ConstantValue rather
than the Constant we make it possible for the owned default value
to be `const static`, ensuring that there's a single copy
of the default for each of sizes and handle rights. Though this
is initially unnecessary (the lifetime of the TypeTemplates is
effectively static), this will not be the case when moving these
default values onto the respective Types as part of
fxbug.dev/74193

fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/517821
Original-Revision: 55927dcb1da792020d362b669a371ac5d01a13da
GitOrigin-RevId: 5bf210d3f6f835bf245671515c8f3dd125302381
Change-Id: Id9dfd68200e16f864af8137f880b644467a01bba