[roll] Roll fuchsia [image assembly] Add support for additional arbitrary HashDescriptors
Adds the ability to create arbitrary HashDescriptors, and a raw builder which
eases the creation of said HashDescriptors.

Adds a field for the minimum required version of avb that a hash descriptor
requires, and when creating a vbmeta image, sets the version based on the maximum
of all required minimum versions.

Original-Bug: 73186

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519985
Original-Revision: cb441bb3e317564de04596060a8e10b2c37ce7d7
GitOrigin-RevId: 4f5747c0761cce797d6f63e03d3667af01cd610f
Change-Id: I2de1814b53f1db531b6baae0513a5a5165c88b10