[roll] Roll infra/recipes [coverage] Keep track of malformed binaries.
There's currently one test that produces a malformed binary due to
fxbug.dev/74189. The covargs tool skips malformed binaries, but we should
keep an allowlist of known tests that have this issue in order to keep
track of them as we work to fix them. fxr/518721 will upload the
mapping of the binary names (which are build ids) to the test labels,
and fxr/519045 will cause covargs to write the list of malformed
binaries to a file. Using these two files, we can compare the malformed
binaries to the list of expected malformed tests. If they don't match,
the build will fail, which means that either a test got fixed or a new
test is malformed, but the check will happen after all the upload steps
so that we can still upload coverage for all other tests.

Original-Bug: 75085
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/519263
Original-Revision: 6fd5b2d65f8c5bcdcffd595098a836202228a0e0
GitOrigin-RevId: 686d43d168a05baf6126becafa61bd3c943ed79d
Change-Id: Ic9e81360427a36f58a5f2a7403592cb804b32678