[roll] Roll fuchsia [kernel] Update JobDispatcher to report exited immediately on job kill
Currently it is possible for userspace to observe zx_info_job_t.exited =
false after a zx_task_kill call on a job if the job has not yet finished
terminating. The documented behavior of zx_info_job_t.exited is that
return_code is valid, which it is even when the job is in
State::KILLING. The current behavior is also inconsistent with
processes, which report exited = true immediately after a zx_task_kill.

This change will assist in debugging ZX_ERR_BAD_STATE errors returned by
fuchsia.process.Launcher that we theorize are from the job that a
process is being launched in terminating in the middle of it being
launched, which is an unavoidable race and not an indication of a deeper
problem.

Original-Bug: 73781
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519666
Original-Revision: 97dd3484ca274191786ac3a6b6f31eea11bbef44
GitOrigin-RevId: f6ffc1816952a8bf811bd0ce5864bd423862ba48
Change-Id: I1d76286d241e720410b6e3bce6f234a4f250939e