[roll] Roll fuchsia [stdcompat]: Add support for is_sorted.
Constexpr is_sorted is a requirement for boot options.
Since we will do the runtime switch to std, we follow the same pattern
as cpp20::is_sorted.
As a side note, added test that verify that the is_sorted behavior
matches the std::is_sorted at runtime.
This verifies that the is_sorted treats < operator, as a <= comparison,
that is:
!comp(a[i+1], a[i]) where comp is the property being checked.

Also updated check to verify that the iterator category is a child of
the required type.

Test: stdcompat-test, stdcompat-unit-test(host)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519507
Original-Revision: 8fea0f8f115e3a96525f45bd44370c3f01e0f881
GitOrigin-RevId: 3589d73602c0d1e2f1eec569a7e8bc45d83d5f73
Change-Id: I6dbe2b0111385200a8e07ee0dd97dafe6998cccf