[roll] Roll fuchsia [dart_tool] Use relative path in metadata
The main problem was that `snapshot` can be a fully qualified GN path,
for example something like `//out/default/...`. Somewhere downstream
this path can be written to a depfile, and it'll be treated as an
absolute path like `/out/default/...`, at least when it goes through
python, which happens to be the problem causing the linked bug.

Also a few other cleanups since I'm here. Note only the line rebasing
`snapshot` in metadata is fixing the bug.

Original-Bug: 75437
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520923
Original-Revision: 8387b32f1525e81b9a0db265980fd3fc3a1715d7
GitOrigin-RevId: d589de227e6b7663b71c26a18eaea6ac6c1fea8f
Change-Id: Iddeaec56811bc83cbd16b5c639dfc9282f0ce690