[roll] Roll fuchsia [zerocopy-derive] Release 0.3.0
Previously, we released a change in 0.2.1 which depends on the
min_const_generics feature, which was stabilized in 1.51. This was a
breaking change for users using compiler versions 1.50 or earlier,
which means it should not have been released as a patch release.

This change bumps the version to 0.3.0. After publishing, I will yank
0.2.1.

This change also releases zerocopy 0.5.0, which depends on
zerocopy-derive 0.3.0, since most users access zerocopy-derive via the
derives' reexports in the zerocopy crate.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521060
Original-Revision: 6c94007b87c6e7cd7808512040ffde4ca525e8d3
GitOrigin-RevId: dc15927125a23a50bda1cef6a4b637780183e3d1
Change-Id: I71bc097f1d57cf2dd8ce7db170007cd450281e3f