[roll] Roll fuchsia [amlogic_decoder] Apply deadline profiles for H264
This CL applies deadline profiles to the various threads used for H264
multi instance decoding in the amlogic decoder driver. This will later
transition to requesting profiles based on the thread's "role" as
described in fxbug.dev/40858, so it introduces the concept of roles
with the expectation that it will be replaced with a different schema
for describing them as part of the aforementioned bug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515382
Original-Revision: 211a11d35ec2559d2362f471d88c748944ea5b54
GitOrigin-RevId: 9f4f2f502695eb9d569f74201f8070153b6efe16
Change-Id: I3748036f4ffb0b95f67cd75d8834134a7d6a5c92