[roll] Roll fuchsia [fint] Fail if no-op check fails
https://fxrev.dev/508218 moved the ninja no-op check from recipes into
fint. In in doing so, it introduced a bug in the ninja no-op check
logic: `fint build` does a ninja dry run and parses the ninja stdio to
check for the "no work to do" string, but then proceeds to ignore the
result, emitting it to the build artifacts but not failing the
subprocess. In turn, the recipe only fails the build if the `fint build`
subprocess has a non-zero retcode.

No-op builds recently became broken as a result of this bug, so this fix
will not be safe to reland until the no-op build is fixed
(https://fxrev.dev/520923).

Original-Bug: 75437
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520581
Original-Revision: 38a64a437d8876b44b45ee4f11ddbd15dfe445b7
GitOrigin-RevId: 4bb19f7b0784423f45333a6bf5e201228d6cae26
Change-Id: I10b51910a70b3964400b5edbcabc1e21ca1822ff