[roll] Roll fuchsia [fint] Emit `ninja explain` stdio in case of no-op failure
Recipes can then read these logs and present their contents to make it
easier for users to debug failed no-op checks. We used to do this when
the no-op check was implemented in recipes, but I removed it for
simplicity when rewriting the no-op check in fint.

Original-Bug: 75437
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521180
Original-Revision: ee3fdc10c937c8084f310b752479528f91d89268
GitOrigin-RevId: 423d2410778f8e1aeae8fe7ef257d4e95826258d
Change-Id: Ia4ca8139df403f7b34d1e317d8b255c6f6fb3738