[roll] Roll fuchsia [fidl][rust] Use write_unaligned in fidl_struct_copy encode
This CL fixes potential undefined behavior in the encoding method for
`fidl_struct_copy!`, which was calling std::ptr::copy_nonoverlapping
a pointer that was non guaranteed to be properly aligned. This replaces
it with a write_unaligned call, which should compile down to the same
thing without any possibility of undefined behavior.

Test: fx test //src/lib/fidl/rust
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521137
Original-Revision: 6ccd99bc04a0fa23c817a922eaf45499805df53f
GitOrigin-RevId: 3a93041665d04ec45ec5f5c2acac164e52c693f1
Change-Id: I00de2bfbc348a163995b022e43ad8c31c6ee3435