[roll] Roll infra/recipes [fint] Emit `ninja explain` stdio in case of no-op failure
Recipes can then read these logs and present their contents to make it
easier for users to debug failed no-op checks. We used to do this when
the no-op check was implemented in recipes, but I removed it for
simplicity when rewriting the no-op check in fint.

Original-Bug: 75437
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521180
GitOrigin-RevId: b70c2f8cb9f2fc7dbe0f510c909e9dfb513b3743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/521021
Original-Revision: 6021c36e2ac62a8e82318c045a5dd39f5afdc4be
Change-Id: I78a7a5351bbbbf4a7b4a566ff239e8370ee664e3