[roll] Roll fuchsia [virtualization][arm64] Improve PL031 emulation.
The PL031 RTC is provided to guests to allow them to determine the wall
time. Currently, Linux is unable to use the PL031 because we fail to
handle reads/writes to the PL031's control register.

This CL updates the our simulated PL031 to support basic reads/writes to
the control register. We also update the update it to more gracefully
handle unknown reads/writes, logging a warning instead of returning an
error (which, in turn, causes the entire VM to crash).

Original-Bug: 74914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520861
Original-Revision: 8af47a0393fe6331841fa9b790268d202f479682
GitOrigin-RevId: 44220d3f9dab86ec50252063134971acff04d7cd
Change-Id: I8f3ba9b6a937460e1ad36d098772ea4778335cfc