[roll] Roll fuchsia [virtualization] Update mklinux.sh to gcc-10.
GCC version 9 is not longer packaged with Debian, making it harder for
users to obtain. Instead, update the script to use GCC 10 which is
actively supported.

We previously did not support to GCC 10, but patches to the Linux kernel
now allow us to successfully build:

  https://zircon-guest-review.googlesource.com/c/third_party/linux/+/2322
  https://zircon-guest-review.googlesource.com/c/third_party/linux/+/2323

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520864
Original-Revision: 185dd3c5c0e1b5aa97d6324982f775464d2fa752
GitOrigin-RevId: 5eb331afb755192bc7bad8a1851e7212e4b7e7f7
Change-Id: Ic16f5d37315d651049facb2c5d5070fe0ab466ca