[roll] Roll fuchsia [virtualization][vmm] Move arch-specific VCPU code into arch/ dir.
Move the architecture-specific HandleMem, HandleInput, HandleOutput,
and HandleIO functions into the arch/ directory. Additionally, prepend
the name "Arch" to the functions to make it a little clearer that these
are per-architecture.

These functions are starting to get a little on the larger side, and the
amount of cross-arch code sharing is minimal. Having per-arch copies
allows us to simplify them, as in the next CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520865
Original-Revision: 7dcd732e908a66c7038bc656ba8530ca4c72ac1e
GitOrigin-RevId: 05033f93d8b24560a161abd3ae3c34756062fe45
Change-Id: I6a64a99fbb2c694a6c32e912e6d9060b0b14d569