[roll] Roll fuchsia [build] Introduce renamed_binary() template.
This CL introduces a new template which will be used in the
few cases where an executable binary must be installed under
a different name inside a package.

The template ensures that the binary is installed to the
right location, _and_ that its runtime libraries are also
installed properly, something that doesn't work with a
simple resource() target.

Moreover, this provides the option of making the renamed
binary visible to the shell, but not the original binary
(which still must be installed). See documentation comments
for details and examples.

This will help in later CL in getting rid of misc. package()
instances (e.g. under //src/sys/pkg/bin/...).

Original-Bug: 65068
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519300
Original-Revision: 9a65eaaa5a619fea934be3dc66b4235e111cbb8c
GitOrigin-RevId: 685b4ff283fd799ec66888a225d3fc455e0b6366
Change-Id: Ide55c34367f92b2929c62fb88990d144a6b2221f