[roll] Roll fuchsia [fidlcat] don't exit if no IR files
If fidlcat can't load the FIDL IR files, it now prints a warning and
allows the execution.

In that case, fidlcat is still able to decode all the syscalls but some
or all the FIDL messages may not be decoded and only dumped in binary.

Original-Bug: 72663
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521281
Original-Revision: 1501bdde608747c9adaaeb7206ac6c71fb8e1a03
GitOrigin-RevId: 3dc714b5b591878cb21f4a07e536fa19b3bc7c6b
Change-Id: Ia9328198f693043ce66bdebbc6efdadc2c8ab4bc