[roll] Roll fuchsia [amlogic_decoder] Apply deadline profiles for VP9
This CL applies deadline profiles to the various threads used for VP9
decoding in the amlogic decoder driver. This will later transition to
requesting profiles based on the thread's "role" as described in
fxbug.dev/40858, so it introduces the concept of roles with the
expectation that it will be replaced with a different schema for
describing them as part of the aforementioned bug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518242
Original-Revision: 11069b5a440f1a54c586c853b145afbc1a69732e
GitOrigin-RevId: e60f3919c97bf44b3ec543b5f7c21f42d7f2eef5
Change-Id: Ife5de53296ea3d9911311b4893cd59b29fa3b273