[roll] Roll fuchsia [kernel] Panic quickly for unexpected page faults in kernel mode
The only kind of page fault that should ever happen in kernel mode
is for a data access during explicit user-copy operations.  For any
other case, a quick panic with full details of the fault site is
crucial to reducing the potential cascade harm of the bug and to
identifying the underlying root cause of the fault.

Original-Bug: 66141
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459941
Original-Revision: 9a97dceec371b1018500b6f527dfa3013e1e6e8e
GitOrigin-RevId: 2b7847726f806fb49c143b952ec1f056c636de04
Change-Id: I5c20b7d233d57b6d508bc3534c1ab901ad7b6572