[roll] Roll fuchsia [nelson][aml-spi] Use deadline profile for Amlogic SPI driver
The next change to the SPI driver will enable interrupt-driven IO. In
order to use interrupts and still meet the timing requirements, a
deadline profile has to be applied to the thread handling the SPI
transfers. This change adds metadata to specify the profile parameters,
and applies the profile in the first transfer request.

Original-Bug: 74692
Test: Let a Nelson running the product config idle for 45 minutes, saw
      no radar driver timeouts. Without this change there were 27
      timeouts in the same interval.
Test: ot-radio driver initialized on Sherlock.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519998
Original-Revision: c904331aff00606027ebc6a1af6443e5318240f7
GitOrigin-RevId: fc9f3ba51254b3334e8b5b72774edd680e84b135
Change-Id: I53f6675cd67fb2e8ebbb52c571c5db8b9cf10739