[roll] Roll fuchsia [zircon][kernel] Validate ZBI dcfg structs and make types portable.
This CL adds size and offset validation to the ZBI driver configuration
structs. As part of this, we add explicit padding to the structs, and
switch to uint8_t instead of bool to make sure that the size is always 1
byte on all platforms.

Original-Bug: 74907
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518925
Original-Revision: 6e9e9cd8caa7910e61a158cb4e23c8dd9e89410d
GitOrigin-RevId: 67d3fd39cafe068956e24721792a8f4a6955a038
Change-Id: I2718a1a0fca319bc2313d6d3edd8d1422a095388