[roll] Roll fuchsia [ffx doctor] match fx spawned daemons with config
As fx often now passes additional flags that are forwarded to daemon
invocations, those need to be accounted for in the pkill pattern.
We're likely approaching the end of the time where we should use
pkill/argument matching for this purpose, though a portable replacement
can be tricky.

Original-Fixed: 75428
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521283
Original-Revision: 86d75062dc5d1b1f31cfd5111f0e0358787ad931
GitOrigin-RevId: 75ebaab2075245591be0e46108c9924720b9a5e2
Change-Id: I66b384a2c506ec45f163f32445d26d7b1f2255c4