[roll] Roll fuchsia [starnix] Organize fs crate
Move FidlFile and SyslogFile to a "fuchsia" subcrate to make room for
other file systems (e.g., procfs, tempfs).

Rename FidlFile to RemoteFile, which matches the terminology we use in
C++ for this concept.

Remove the Mutex around the SynchronousProxy because SynchronousProxy is
not thread-safe.

Rename FileDesc to FileObject because FileDesc was ambiguous between
FileDescriptor and FileDescription.

Rename FdNumber to FileDescriptor because that is the common industry
term for this type.

Rename FdHandle to FileHandle because its a handle to the actual file
rather than a handle to the file descriptor.

Finally, make the i32 inside FileDescriptor private. There is already a
raw() accessor for it.  We do not need to leak the structure of the type
itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521906
Original-Revision: 36a6ef17c38504948ac3ece4952cc5ccd85e7352
GitOrigin-RevId: b8c408937417df4d7c8512482bdba2a3955afc9d
Change-Id: I8c4ddd15804e467e65f8bb69ea13c1d4b6ddae71