[roll] Roll fuchsia [scripts][clang] Add script for checking that vtables are in rodata
This is the script I used for checking if all vtables in a file, fuchsia
toolchain, llvm multilib, or fuchsia checkout are in the .rodata
section. The script works by naively parsing `readelf` output, finding
vtable symbols, and checking if they're in .rodata. It could be there's
a more production-friendly way of checking, but this is just what I
whipped up for local testing.

This is intended to be used by the clang canary builders as a final
check to ensure that vtables under the relative vtables ABI stay
readonly.

Original-Bug: 45314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513859
Original-Revision: d09cbd06294ad21f877e7565faeae693daf1ce7c
GitOrigin-RevId: a4dbf7e4f43a5899c486b441bb08620189ff98cd
Change-Id: Iaa506b3b2f17f57f46f2dead08f4f427fef356d3