[roll] Roll fuchsia [starnix] Watch for exceptions on each thread
Previously, we watched for syscalls on the process object, but that
makes it difficult to create additional threads in the process. Now, we
watch for syscalls on the thread object, which means we can have a
thread in starnix for each thread in the child process once we create
multiple threads in a child process.

Test: No behavior change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522201
Original-Revision: 8a388aa6e00214aaad7767ea808f0be24cdf5b03
GitOrigin-RevId: babfff1cb5082d16e1fe8a2c35067933cb09da3b
Change-Id: Ie5b091f21c881f942f77e64a62b2f6a4d66bb4c4