[roll] Roll fuchsia [modular] Fix crash when presenting a pending view for a deleted story
When a story is created before sessionmgr chooses a presentation
protocol, i.e. before the session shell's outgoing directory is
served, the story's view is pended, and the view is presented
later when the protocol is chosen.

If the client deleted the story before sessionmgr connected to
the the presentation protocol, sessionmgr would try to read a
nonexistent StoryData and crash.

Original-Fixed: 73335
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520125
Original-Revision: 6c1996ee8cac4c0e75451d0434a3cf84b7891a20
GitOrigin-RevId: fba88d3b3d249bc72b90a23c6ad59aa95a75417f
Change-Id: Id5126e48c8eb3a6467efb0d12ec240006d67435d