[roll] Roll fuchsia [fidlc] Move layout arg resolution into TypeTemplate
This CL moves layout parameter and type constraint resolution
code from Library::CompileTypeConstructor into
TypeTemplate::ResolveArgs. As a side effect, the data that
is piped through

  Typespace::Create -> Typespace::CreateNotOwned ->
  TypeTemplate::Create

now consists of unresolved arguments
(wrapped in an ArgsAndConstraints struct). The previous
CreateInvocation still remains, but is now an output of
TypeTemplate::ResolveArgs, which various TypeTemplates must
call in order to resolve their arguments.

Standalone, this CL introduces complexity since now every
TypeTemplate needs to call ResolveArgs, instead of having this
done in a single place before calling into the typespace, but
it sets up the infrastructure needed to resolve parameters/
constraints on a per-TypeTemplate basis, which is necessary
in the new syntax. Another valid option would have been
to keep this the same for the old syntax and have a different
flow altogether for the new syntax but the consistency should
help with making the code easier to understand

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64-asan/exe.unstripped/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/513942
Original-Revision: eda7e447cea16fe53f43caf5ecb120b787e0caa4
GitOrigin-RevId: b9b8c9bd13a19267df163f66d8129b5cb09dea16
Change-Id: I44b09a275661d8810f43fa457d1216e219255f34