[roll] Roll fuchsia [starnix] Implement getppid, getpgid, and getpgrp
Implementing these functions lead to a refactoring of how tasks are
owned. After this CL, all the task-related data-structures are owned
from the "leaves" of the tree, which are the task objects themselves.

To make the ownership of Task objects clear, Task::new now returns a
TaskOwner object, which has a single owner. When that owner drops the
TaskOwner, the Task object is removed from the ThreadGroup and the
PidTable. When all the Tasks in a ThreadGroup have been removed, the
ThreadGroup kills the underlying Zircon process.

Original-Bug: 74723
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/521864
Original-Revision: 2d8009168dd0eea28959184be6a937abf812d64f
GitOrigin-RevId: d94be847f3758102b08ae98a5a2720aea041c386
Change-Id: Ic7dfd53b299fb4d95fb744474ea71dc2722ab01a