[roll] Roll fuchsia [gidl][llcpp] Special case LLCPP byte vector build benchmarks
The existing benchmarks won't compile quickly for sizes >4k
on arm with clang.

This change adds logic that detects simple repeating patterns
and uses memcpy to reduce the number of operations the
compiler needs to deal with. This was proposed previously,
but I was avoiding doing it because it seemed hacky. Despite
the hackiness, I think this will be useful for benchmarking
iovec encoding.

Original-Bug: 66977
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520360
Original-Revision: 7cd4b385bcf404d579ecd790c54810e5944d7590
GitOrigin-RevId: 7ab086ccf66660b72c8a9176c8bf881eec1fe54f
Change-Id: Ia6d5014f6948915f5a6859e4cb6a7311bc699293