[roll] Roll fuchsia [bt][gap] Reduce log severity on errors caused by client disconnect.
A client is allowed to disconnect at anytime. It is not an error to
discover that a client has disconnected when attempting to send it
FIDL messages. Log severity reduced to INFO since it could be useful in
understanding exactly what messages bt-gap believes it sent to a client.

Original-Bug: 75806
Test: Log severity change only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522670
Original-Revision: 8d7fa3f83dc2a83e0c47e39b128c67e22366ba20
GitOrigin-RevId: 6bf41319ba723f0dfe907dc55b1cedba07af2ba8
Change-Id: I4e03b85f0f940c9dc186a26c8ea0723e5f1962b6