[roll] Roll fuchsia [fidlc] Move declaration sorting to post-compilation.
Declaration sorting depends on certain information about the
types being sorted, like the nullability of a type, to determine
the dependencies of a type. In the old syntax we have the
information we need before needing to compile any types,
but this is not the case in the new syntax, so as preparation
for compiling "new syntax" types this CL moves declaration
sorting to be after type compilation.

This also updates the type dependency logic to use Types as
the source of information whenever possible, rather than the
TypeConstructor, which allows us to avoid doing things like
doing a string comparison against "request" to see if a type
refers to a request handle (though currently this hard coding
is OK because fidlc fails to shadow built-in types anyways).

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64-asan/exe.unstripped/fidl-compiler

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/518482
Original-Revision: 56b72ad27145d00b2ba14b7a365655ab3fd608a7
GitOrigin-RevId: c4d46ab102a9afc37dd148b8868b4071d8024956
Change-Id: Ibbe2ce47b189d9d8eb7668138b5be8c72e1e48f8