[roll] Roll fuchsia [build][image assembly] use ffx assembly plugin to generate vbmeta image
This adds actions to all builds with vbmeta image generation to
additionally use the new ffx plugin to generate a vbmeta image, and then
an additional action to compare that the two files are the same
(that they have identical contents).

This is a preliminary step to swapping which action's output is used by
the latter steps (while retaining the generation of the file by avbtool)
and then a follow-on change will remove the generate of the file by
avbtool entirely.

Original-Bug: 73187
Test: Lots of CQ dry-runs, with in-build verification step

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519986
Original-Revision: 481389e628987e99d7e2a202b3e40650c5113faf
GitOrigin-RevId: 973f25a3662d9e4ebd08e799e56ea302199b428a
Change-Id: I2cd7432b156946d48f1f7c1702cd02391040f994