[roll] Roll infra/recipes [coverage] Delete "checkout_root" output property
It's no longer load-bearing since fxrev.dev/500618 landed.

It used to be necessary to work around a bbagent bug/feature that
randomized build start directories, because covargs operates on absolute
paths. But that bug has been fixed and it may no longer even be the case
that covargs operates on absolute paths, which would make it extremely
unlikely that we'd have to reintroduce this feature in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/521724
Original-Revision: d67d2508694d0dc00226fcd67279a7bbf3b4edce
GitOrigin-RevId: 919133a8c29ccf7b9f7cfaa0185a0fbd3021107d
Change-Id: Id532a3422187cac9e0631cea90e0ef4938bf4bb9